% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction_page.R
\docType{methods}
\name{add_dim_reduction_page,i2dashboard,SingleCellExperiment-method}
\alias{add_dim_reduction_page,i2dashboard,SingleCellExperiment-method}
\title{Renders a page with four linked components at maximum. The first component is a scatterplot, showing samples in along coordinates from \code{reduced_dim}. If \code{barplot_grouping} is \code{TRUE} the second component will be a bar plot, that shows the number of observations by groups defined in \code{grouping}. If \code{silhouette_plot} is \code{TRUE} the third component will be a silhoutte plot, that shows how well each observation has been classified. Therefore the Euclidean distance is applied to the coordinates from \code{reduced_dim} and grouped by the groups defined in \code{grouping}. The remaining component slots are filled with bar or boxplots, that visualize one of the columns from \code{metadata} by groups defined in \code{grouping}.}
\usage{

  \S4method{add_dim_reduction_page}{i2dashboard,SingleCellExperiment}(object,
  sc_object, reduced_dim, metadata, grouping, title = NULL,
  barplot_grouping = TRUE, silhouette_plot = FALSE, menu = NULL,
  sidebar = NULL)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{sc_object}{A valid \linkS4class{SingleCellExperiment::SingleCellExperiment} object}

\item{reduced_dim}{A character of length 1 representing the name of a dimension reduction of \code{reducedDim}.}

\item{metadata}{A character or list representing the names of  columns of \code{colData}.}

\item{grouping}{A character of length 1 representing the name of columns of \code{colData} and used for expression grouping.}

\item{title}{The title of the page.}

\item{barplot_grouping}{(Optional) A logical value. (Default: TRUE) If TRUE, a barplot with the number of observations from 'grouping' will be created.}

\item{silhouette_plot}{(Optional) A logical value. (Default: FALSE) If TRUE, a silhouette plot will be created.}

\item{menu}{(Optional) The name of the menu, under which the page should appear in the navigation.}

\item{sidebar}{(Optional) The page layout (see below).}
}
\value{
A string containing markdown code for the rendered page.
}
\description{
Renders a page with four linked components at maximum. The first component is a scatterplot, showing samples in along coordinates from \code{reduced_dim}. If \code{barplot_grouping} is \code{TRUE} the second component will be a bar plot, that shows the number of observations by groups defined in \code{grouping}. If \code{silhouette_plot} is \code{TRUE} the third component will be a silhoutte plot, that shows how well each observation has been classified. Therefore the Euclidean distance is applied to the coordinates from \code{reduced_dim} and grouped by the groups defined in \code{grouping}. The remaining component slots are filled with bar or boxplots, that visualize one of the columns from \code{metadata} by groups defined in \code{grouping}.
}
