% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction_page.R
\name{add_dim_reduction_page}
\alias{add_dim_reduction_page}
\title{Renders a dimension reduction page}
\usage{
add_dim_reduction_page(object, reduced_dim, metadata, grouping,
  title = NULL, labels = NULL, barplot_grouping = TRUE,
  silhouette_plot = FALSE, menu = NULL, sidebar = NULL)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{reduced_dim}{A data.frame or matrix containing two columns for the x and y coordinates.}

\item{metadata}{A data.frame or matrix containing a maximum of four columns with cell metadata (e.g. cluster, timepoint, number of genes, etc). Should have the same rownumber as 'reduced_dim'. Note: Factorial metadata needs to be of class factor in the data.frame.}

\item{grouping}{A character string identical to one of the column names in 'metadata'.}

\item{title}{The title of the page.}

\item{labels}{(Optional) A vector with labels. Should be of the same length as the rownumber of 'reduced_dim'.}

\item{barplot_grouping}{(Default: TRUE) A logical value. (Default: TRUE) If TRUE, a barplot with the number of observations from 'grouping' will be created.}

\item{silhouette_plot}{(Optional) A logical value. (Default: FALSE) If TRUE, a silhouette plot will be created.}

\item{menu}{(Optional) The name of the menu, under which the page should appear in the navigation.}

\item{sidebar}{(Optional) The page layout (see below).}
}
\value{
A string containing markdown code for the rendered page.
}
\description{
Renders a dimension reduction page
}
