% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred_metadata_page.R
\docType{methods}
\name{add_dimred_metadata_page}
\alias{add_dimred_metadata_page}
\alias{add_dimred_metadata_page,i2dashboard,missing-method}
\alias{add_dimred_metadata_page,i2dashboard,SingleCellExperiment-method}
\alias{add_dimred_metadata_page,i2dashboard,Seurat-method}
\title{Add a dimension reduction with metadata page}
\usage{
add_dimred_metadata_page(report, object, ...)

\S4method{add_dimred_metadata_page}{i2dashboard,missing}(report,
  use_dimred, exprs_values, metadata, title = "Marker gene expression",
  menu = NULL)


  \S4method{add_dimred_metadata_page}{i2dashboard,SingleCellExperiment}(report,
  object, use_dimred, exprs_values, metadata, genes,
  title = "Marker gene expression", menu = NULL)

\S4method{add_dimred_metadata_page}{i2dashboard,Seurat}(report, object,
  use_dimred, exprs_values, metadata, genes, assay, assay_slot = "data",
  title = "Marker gene expression", menu = NULL)
}
\arguments{
\item{report}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{A valid \linkS4class{SingleCellExperiment::SingleCellExperiment} object.}

\item{use_dimred}{A data.frame or matrix containing coordinates of the reduced dimensions.}

\item{exprs_values}{A data.frame or matrix containing expression data of features of interest (features in rows and samples in columns). Rownames are used as feature names.}

\item{metadata}{A data.frame or matrix along rows of \code{exprs_values} containing metadata data for the features of interest.}

\item{title}{The title of the page.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}

\item{assay}{A character of length 1 representing the name of a \linkS4class{Seurat::Assay} object present in the slot \code{assays} of the \linkS4class{Seurat::Seurat} object.}

\item{assay_slot}{(Default value: "data") A character of length 1 representing the name of a slot in the \linkS4class{Seurat::Assay} object provided in \code{assay}.}

\item{object}{A valid \linkS4class{Seurat::Seurat} object.}
}
\value{
A string containing markdown code for the rendered page.

A string containing markdown code for the rendered page.

A string containing markdown code for the rendered page.
}
\description{
This function renders adds a page with two linked components to the \code{report} object: The first component is a scatterplot, showing samples in along two-dimensional coordinates from \code{use_dimred}. The second component is a table, that shows metadata from \code{metadata}. The scatterplot is colored by feature expression values provided in \code{expression} for the selected feature in the table.
}
