% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/feature_expression_page.R
\name{add_feature_expression_page}
\alias{add_feature_expression_page}
\alias{add_feature_expression_page,i2dashboard,missing-method}
\alias{add_feature_expression_page,i2dashboard,SingleCellExperiment-method}
\alias{add_feature_expression_page,i2dashboard,Seurat-method}
\title{Add a feature expression page.}
\usage{
add_feature_expression_page(dashboard, object, ...)

\S4method{add_feature_expression_page}{i2dashboard,missing}(
  dashboard,
  use_dimred,
  exprs_values,
  group_by = NULL,
  labels = rownames(use_dimred),
  title = "Feature expression",
  menu = NULL
)

\S4method{add_feature_expression_page}{i2dashboard,SingleCellExperiment}(
  dashboard,
  object,
  use_dimred,
  exprs_values,
  group_by = NULL,
  subset_row = NULL,
  title = "Feature expression",
  menu = NULL
)

\S4method{add_feature_expression_page}{i2dashboard,Seurat}(
  dashboard,
  object,
  use_dimred,
  assay,
  group_by,
  slot = "data",
  subset_row = NULL,
  title = "Feature expression",
  menu = NULL
)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{A valid \linkS4class{SingleCellExperiment::SingleCellExperiment} or \linkS4class{Seurat::Seurat} object.}

\item{use_dimred}{Coordinates of the reduced dimensions, used for the scatterplot (see Details).}

\item{exprs_values}{Expression data of features of interest in rows and samples in columns (see Details).}

\item{group_by}{Data along samples that is used for grouping expression values in the violin plot (see Details).}

\item{labels}{A vector with optional sample labels that are used instead of \code{rownames(use_dimred)}.}

\item{title}{The title of the page.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}

\item{subset_row}{A character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use. The default of NULL will use all features.}

\item{assay}{A character vector specifying which assay from \code{object@assays} to obtain expression values from (see Details).}

\item{slot}{A character vector specifying the name of the slot in the assay.}

\item{page}{The name of the page to be added.}

\item{assay_slot}{A character specifying the name of the data slot in the assay. (Default: "data")}
}
\value{
An object of class \linkS4class{i2dash::i2dashboard}.

An object of class \linkS4class{i2dash::i2dashboard}.

An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
This function adds a page with two linked components to the \code{dashboard} object: A scatterplot, showing samples in along two-dimensional coordinates, and a violin plot, showing feature expression values by groups defined in \code{group_by}.
}
\details{
The parameters \code{use_dimred}, \code{exprs_values} (or \code{assay}) and \code{group_by} take different arguments depending on the class of \code{object}.
  In case no object is supplied (\emph{i2dashboard,missing}-method), the parameters are expected to be of class \code{data.frame} or \code{matrix}.
  In case a \linkS4class{SingleCellExperiment::SingleCellExperiment} or \linkS4class{Seurat::Seurat} object is supplied, the parameters are expected to be of class \code{character}, containing
  \itemize{
    \item the name of an item in \code{reducedDims(object)} or \code{object@reductions},
    \item a valid assay name from \code{assayNames(object)} or \code{names(object@assays)},
    \item column names of \code{colData(object)} or \code{object@meta.data}.
  }
}
