% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression_page.R
\docType{methods}
\name{add_gene_expression_page,i2dashboard,Seurat-method}
\alias{add_gene_expression_page,i2dashboard,Seurat-method}
\title{Renders a page with two linked components. The first component is a scatterplot, showing samples in along coordinates from \code{reduced_dim}. The second component is a violin plot, that shows expression values from \code{expression} by groups defined in \code{grouping}.}
\usage{
\S4method{add_gene_expression_page}{i2dashboard,Seurat}(object, sc_object,
  reduced_dim, metadata, grouping, assay, assay_slot = "data", genes,
  title = NULL, menu = NULL, sidebar = NULL)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{sc_object}{A valid \linkS4class{Seurat::Seurat} object}

\item{reduced_dim}{A character of length 1 representing the name of a \linkS4class{Seurat::DimReduc} object present in the slot \code{reductions} of the \linkS4class{Seurat::Seurat} object.}

\item{metadata}{A character or list representing the names of columns in the slot \code{meta.data} in the \linkS4class{Seurat::Seurat} object.}

\item{grouping}{A character of length 1 representing the name of a column in the slot \code{meta.data} in the \linkS4class{Seurat::Seurat} and used for expression grouping.}

\item{assay}{A character of length 1 representing the name of a \linkS4class{Seurat::Assay} object present in the slot \code{assays} of the \linkS4class{Seurat::Seurat} object.}

\item{assay_slot}{(Default value: "data") A character of length 1 representing the name of a slot in the \linkS4class{Seurat::Assay} object provided in \code{assay}.}

\item{genes}{A character or list of genes of interest that are present in row names of the provided \code{'assay_slot'}.}

\item{title}{The title of the page.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}

\item{sidebar}{(Optional) The page layout (see below).}
}
\value{
A string containing markdown code for the rendered page.
}
\description{
Renders a page with two linked components. The first component is a scatterplot, showing samples in along coordinates from \code{reduced_dim}. The second component is a violin plot, that shows expression values from \code{expression} by groups defined in \code{grouping}.
}
