% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression_page.R
\docType{methods}
\name{add_gene_expression_page,i2dashboard,SingleCellExperiment-method}
\alias{add_gene_expression_page,i2dashboard,SingleCellExperiment-method}
\title{Renders a page with two linked components. The first component is a scatterplot, showing samples in along coordinates from \code{reduced_dim}. The second component is a violin plot, that shows expression values from \code{expression} by groups defined in \code{grouping}.}
\usage{

  \S4method{add_gene_expression_page}{i2dashboard,SingleCellExperiment}(object,
  sc_object, reduced_dim, expression, metadata, grouping, genes,
  title = NULL, menu = NULL, sidebar = NULL)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{sc_object}{A valid \linkS4class{SingleCellExperiment::SingleCellExperiment} object}

\item{reduced_dim}{A character of length 1 representing the name of a dimension reduction of \code{reducedDim}.}

\item{expression}{A character of length 1 representing the name of an \code{assay}.}

\item{metadata}{A character or list representing the names of  columns of \code{colData}.}

\item{grouping}{A character of length 1 representing the name of columns of \code{colData} and used for expression grouping.}

\item{genes}{A character or list of genes of interest that are present in row names of the \code{reduced_dim}.}

\item{title}{The title of the page.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}

\item{sidebar}{(Optional) The page layout (see below).}
}
\value{
A string containing markdown code for the rendered page.
}
\description{
Renders a page with two linked components. The first component is a scatterplot, showing samples in along coordinates from \code{reduced_dim}. The second component is a violin plot, that shows expression values from \code{expression} by groups defined in \code{grouping}.
}
