% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression_page.R
\docType{methods}
\name{add_gene_expression_page}
\alias{add_gene_expression_page}
\alias{add_gene_expression_page,i2dashboard,missing-method}
\alias{add_gene_expression_page,i2dashboard,SingleCellExperiment-method}
\title{Add a gene expression page.}
\usage{
add_gene_expression_page(report, object, ...)

\S4method{add_gene_expression_page}{i2dashboard,missing}(report,
  use_dimred, exprs_values, group_by = NULL,
  labels = rownames(use_dimred), title = "Gene expression",
  menu = NULL)


  \S4method{add_gene_expression_page}{i2dashboard,SingleCellExperiment}(report,
  object, use_dimred, expression, metadata, grouping, genes,
  title = NULL, menu = NULL, sidebar = NULL)
}
\arguments{
\item{report}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{A valid \linkS4class{SingleCellExperiment::SingleCellExperiment} object.}

\item{use_dimred}{A data.frame or matrix containing coordinates of the reduced dimensions.}

\item{exprs_values}{A data.frame or matrix containing expression data of features of interest in rows and samples in columns.}

\item{group_by}{A factor or data.frame along rows of \code{use_dimred} that is used for grouping expression values in the violin plot.}

\item{labels}{A vector with optional sample labels that are used instead of \code{rownames(use_dimred)}.}

\item{title}{The title of the page.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}
}
\value{
A string containing markdown code for the rendered page.

A string containing markdown code for the rendered page.
}
\description{
This function adds a page with two linked components to the \code{report} object: A scatterplot, showing samples in along two-dimensional coordinates, and a violin plot, showing feature expression values by groups defined in \code{group_by}.
}
