% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_gene_expression_page.R
\docType{methods}
\name{add_gene_grid_page,i2dashboard,missing-method}
\alias{add_gene_grid_page,i2dashboard,missing-method}
\title{Renders a page, where the user first can select the dimension reduction from \code{reduced_dim} (if a list is given). Then several features from \code{expression} can be selected and also the number of columns to render scatter plots along coordinates from \code{reduced_dim} for each selected feature and coloured by its expression values from \code{expression}.}
\usage{
\S4method{add_gene_grid_page}{i2dashboard,missing}(object, reduced_dim,
  expression, title = NULL, menu = NULL)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{reduced_dim}{A named list of data.frames or matrices or a single data.frame or matrix containing coordinates of the reduced dimensions.}

\item{expression}{A data.frame or matrix containing expression data of features of interest in rows and samples in columns.}

\item{title}{The title of the page.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}
}
\value{
A string containing markdown code for the rendered page.
}
\description{
Renders a page, where the user first can select the dimension reduction from \code{reduced_dim} (if a list is given). Then several features from \code{expression} can be selected and also the number of columns to render scatter plots along coordinates from \code{reduced_dim} for each selected feature and coloured by its expression values from \code{expression}.
}
