% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_gene_expression_page.R
\docType{methods}
\name{add_gene_grid_page,i2dashboard,Seurat-method}
\alias{add_gene_grid_page,i2dashboard,Seurat-method}
\title{Renders a page, where the user first can select the dimension reduction from \code{reduced_dim} (if a list is given). Then several features from \code{expression} can be selected and also the number of columns to render scatter plots along coordinates from \code{reduced_dim} for each selected feature and coloured by its expression values from \code{expression}.}
\usage{
\S4method{add_gene_grid_page}{i2dashboard,Seurat}(object, sc_object,
  reduced_dim, assay, assay_slot = "data", genes, title = NULL,
  menu = NULL)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{sc_object}{A valid \linkS4class{Seurat::Seurat} object}

\item{reduced_dim}{reduced_dim A character of length 1 representing the name of a \linkS4class{Seurat::DimReduc} object present in the slot \code{reductions} of the \linkS4class{Seurat::Seurat} object.}

\item{assay}{A character of length 1 representing the name of a \linkS4class{Seurat::Assay} object present in the slot \code{assays} of the \linkS4class{Seurat::Seurat} object.}

\item{assay_slot}{(Default value: "data") A character of length 1 representing the name of a slot in the \linkS4class{Seurat::Assay} object provided in \code{assay}.}

\item{genes}{A character or list of genes of interest that are present in row names of the provided \code{'assay_slot'}.}

\item{title}{The title of the page.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}
}
\value{
A string containing markdown code for the rendered page.
}
\description{
Renders a page, where the user first can select the dimension reduction from \code{reduced_dim} (if a list is given). Then several features from \code{expression} can be selected and also the number of columns to render scatter plots along coordinates from \code{reduced_dim} for each selected feature and coloured by its expression values from \code{expression}.
}
