% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_gene_expression_page.R
\docType{methods}
\name{add_gene_grid_page,i2dashboard,SingleCellExperiment-method}
\alias{add_gene_grid_page,i2dashboard,SingleCellExperiment-method}
\title{Renders a page, where the user first can select the dimension reduction from \code{reduced_dim} (if a list is given). Then several features from \code{expression} can be selected and also the number of columns to render scatter plots along coordinates from \code{reduced_dim} for each selected feature and coloured by its expression values from \code{expression}.}
\usage{
\S4method{add_gene_grid_page}{i2dashboard,SingleCellExperiment}(object,
  sc_object, reduced_dim, expression, genes, title = NULL, menu = NULL)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{sc_object}{A valid \linkS4class{SingleCellExperiment::SingleCellExperiment} object}

\item{reduced_dim}{A list or character representing the name of a dimension reduction of \code{reducedDim} in the SingleCellObject.}

\item{expression}{A character of length 1 representing the name of an \code{assay}.}

\item{genes}{A character or list of genes of interest that are present in row names of the \code{reduced_dim}.}

\item{title}{The title of the page.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}
}
\value{
A string containing markdown code for the rendered page.
}
\description{
Renders a page, where the user first can select the dimension reduction from \code{reduced_dim} (if a list is given). Then several features from \code{expression} can be selected and also the number of columns to render scatter plots along coordinates from \code{reduced_dim} for each selected feature and coloured by its expression values from \code{expression}.
}
