% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_gene_expression_page.R
\name{add_multi_gene_expression_page}
\alias{add_multi_gene_expression_page}
\title{Renders a page with two linked components. The first component is a scatterplot, showing samples in along coordinates from \code{reduced_dim}. The second component is a violin plot, that shows expression values from \code{expression} by groups defined in \code{grouping}.}
\usage{
add_multi_gene_expression_page(object, reduced_dim, expression,
  title = NULL, menu = NULL)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{reduced_dim}{A named list of data.frames or matrices or a single data.frame or matrix containing coordinates of the reduced dimensions.}

\item{expression}{A data.frame or matrix containing expression data of features of interest in rows and samples in columns.}

\item{title}{The title of the page.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}
}
\value{
A string containing markdown code for the rendered page.
}
\description{
Renders a page with two linked components. The first component is a scatterplot, showing samples in along coordinates from \code{reduced_dim}. The second component is a violin plot, that shows expression values from \code{expression} by groups defined in \code{grouping}.
}
