% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_characterization_barplot.R
\name{cluster_characterization_barplot}
\alias{cluster_characterization_barplot}
\title{Renders a barplot for cluster characterization}
\usage{
cluster_characterization_barplot(object, x, cluster,
  title = "Characterization of clusters")
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{x}{Numeric values mapped to the x-axis. In case of a nested list, a dropdown menu will be provided in the interactive mode. If x is NULL then a barplot for "Number of cells" per cluster will be created. If x is not NULL a barplot for "Fraction per cell" per cluster will be created.}

\item{cluster}{Values for the membership to clusters. In case of a nested list, a dropdown menu will be provided in the interactive mode.}

\item{title}{A title that will be displayed on top.}
}
\value{
A string containing markdown code for the rendered textbox
}
\description{
Renders a barplot for cluster characterization
}
