% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{ComplexHeatmap_heatmap}
\alias{ComplexHeatmap_heatmap}
\title{Render a heatmap with ComplexHeatmap.}
\usage{
ComplexHeatmap_heatmap(..., legend_title = NULL)
}
\arguments{
\item{...}{Additional arguments passed on to \code{ComplexHeatmap::Heatmap}.}

\item{legend_title}{An optional title of the legend.}
}
\value{
An object of class \code{Heatmap}.
}
\description{
Render a heatmap with ComplexHeatmap.
}
