% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{correct_name}
\alias{correct_name}
\title{Helper function to correct column names
See https://github.com/tidyverse/dplyr/issues/1902 for additional information.}
\usage{
correct_name(key, prefix)
}
\arguments{
\item{key}{The key.}

\item{prefix}{The prefix.}
}
\value{
The corrected column names.
}
\description{
Helper function to correct column names
See https://github.com/tidyverse/dplyr/issues/1902 for additional information.
}
