% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{create_barplot_df}
\alias{create_barplot_df}
\title{Function to create a dataframe for plotly_barplot.}
\usage{
create_barplot_df(group_by, x = NULL)
}
\arguments{
\item{group_by}{A list with factorial values, by which observations can optionally be grouped.}

\item{x}{(Optional) A named list with the observations for the barplot.}
}
\value{
An object of class \code{list} containig the dataframe 'df', the vector 'x' with values for the x-axis, the vector 'y' with values for the y-axis, the vector 'names' (can be NULL), the boolean value 'showlegend'.
}
\description{
Function to create a dataframe for plotly_barplot.
}
