% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{create_violinplot_df}
\alias{create_violinplot_df}
\title{Function to create a dataframe for plotly_violinplot.}
\usage{
create_violinplot_df(y, group_by = NULL, title_y = NULL,
  title_group_by = NULL)
}
\arguments{
\item{y}{A list with the observations for the violinplot.}

\item{group_by}{(Optional) A list with factorial values, by which observations can optionally be grouped.}

\item{title_y}{(Optional) The title of the y-axis that describes the observations.}

\item{title_group_by}{(Optional) The title of the x-axis that describes the grouping factor.}
}
\value{
An object of class \code{list} containig the dataframe 'df', the vector 'x' with values for the x-axis, the vector 'y' with values for the y-axis, the vector 'split', the boolean value 'showlegend', the character string "title_y", the character string "title_group_by".
}
\description{
Function to create a dataframe for plotly_violinplot.
}
