% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{df_per_func}
\alias{df_per_func}
\title{Function to create a data.frame per function: rownames = selected colData | column for each level of group_by}
\usage{
df_per_func(data, columns, func, group_by)
}
\arguments{
\item{data}{A data.frame.}

\item{columns}{The columns of \code{data} that are summarized.}

\item{func}{A character of one summary function.}

\item{group_by}{A vector of the same length as the number of rows in data, which is used for grouping.}
}
\value{
data.frame
}
\description{
Function to create a data.frame per function: rownames = selected colData | column for each level of group_by
}
