% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{df_without_grouping}
\alias{df_without_grouping}
\title{Function to create a data.frame containing one column for each provided function (rownames = selected colData)}
\usage{
df_without_grouping(data, columns, FUNS)
}
\arguments{
\item{data}{A data.frame.}

\item{columns}{The columns of \code{data} that are summarized.}

\item{FUNS}{A character of one summary function.}
}
\value{
data.frame
}
\description{
Function to create a data.frame containing one column for each provided function (rownames = selected colData)
}
