% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dimred_feature_page.R
\docType{methods}
\name{dimred-feature-page}
\alias{dimred-feature-page}
\alias{add_dimred_feature_page}
\title{View a dimension reduction side-by-side with feature metadata}
\usage{
add_dimred_feature_page(dashboard, object, ...)

\S4method{add_dimred_feature_page}{i2dashboard,missing}(dashboard,
  use_dimred, exprs_values, feature_metadata,
  page = "dimred_feature_page", title = "Feature expression",
  menu = NULL)


  \S4method{add_dimred_feature_page}{i2dashboard,SingleCellExperiment}(dashboard,
  object, use_dimred, exprs_values, feature_metadata, subset_row, ...)

\S4method{add_dimred_feature_page}{i2dashboard,Seurat}(dashboard, object,
  use_dimred, feature_metadata, subset_row, assay = "RNA",
  assay_slot = "data", ...)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{A \linkS4class{SingleCellExperiment::SingleCellExperiment} object or a \linkS4class{Seurat::Seurat} object.}

\item{use_dimred}{A data.frame (matrix) containing coordinates of the reduced dimension or a character representing a valid \code{reducedDim} slot of a SingleCellExperiment \code{object}/ a name of the \linkS4class{Seurat::DimReduc} object in \code{object@reductions}.}

\item{exprs_values}{A data.frame (matrix) containing expression data of features of interest in rows and samples in columns.}

\item{feature_metadata}{A data.frame (matrix) along rows of \code{exprs_values} containing feature metadata, or a character vector indicating columns from \code{rowData(object)} or \code{object[[assay]]@feature.data}.}

\item{page}{The name of the page to be added.}

\item{title}{The title of the page.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}

\item{assay}{A character specifying the assay (\code{object@assays}) to obtain expression values from. (Default: "RNA")}

\item{assay_slot}{A character specifying the name of the data slot in the assay. (Default: "data")}
}
\value{
An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
The dimension reduction plot is colored by feature expression and updated if the users clicks feature rows in the metadata table.
}
