% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{.bar_plot}
\alias{.bar_plot}
\title{Render a bar plot}
\usage{
.bar_plot(cluster, x = NULL)
}
\arguments{
\item{cluster}{Values for the membership to clusters. In case of a nested list, a dropdown menu will be provided in the interactive mode.}

\item{x}{Numeric values mapped to the x-axis. In case of a nested list, a dropdown menu will be provided in the interactive mode.}
}
\value{
A list with 1. the plotly object & 2. the data frame used in the plot
}
\description{
Render a bar plot
}
