% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{.scatter_plot}
\alias{.scatter_plot}
\title{Render a scatter plot}
\usage{
.scatter_plot(labels = NULL, x, y, colour_by = NULL,
  checkbox = FALSE, selected_label = NULL)
}
\arguments{
\item{labels}{A list with sample names, that should be of the same length as x and y.}

\item{x}{Numeric values mapped to the x-axis. In case of a nested list, a dropdown menu will be provided in the interactive mode.}

\item{y}{Numeric values mapped to the y-axis. In case of a nested list, a dropdown menu will be provided in the interactive mode.}

\item{colour_by}{A factor that will be mapped to colours. In case of a nested list, a dropdown menu will be provided in the interactive mode.}
}
\value{
A list with 1. the plotly object & 2. the data frame used in the plot
}
\description{
Render a scatter plot
}
