% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_discarded.R
\name{.summarize_discarded}
\alias{.summarize_discarded}
\title{Summarizes columns of \code{data} by \code{discarded} and is grouped by \code{columns}.}
\usage{
.summarize_discarded(
  data,
  discarded,
  columns,
  colnames,
  table_title,
  footnote_title,
  footnote_text
)
}
\arguments{
\item{data}{A data.frame.}

\item{discarded}{A column of \code{data} contatining logical values.}

\item{columns}{The columns of \code{data} containing experiemntal factors.}

\item{colnames}{A vector of length three containing the column names of the final table.}

\item{table_title}{Title of the table.}

\item{footnote_title}{Title of the footnote.}

\item{footnote_text}{Description for the table shown as footnote.}
}
\value{
A list with a \code{data.frame} and a HTML table that contains the number of kept and disarded samples per column.
}
\description{
Summarizes columns of \code{data} by \code{discarded} and is grouped by \code{columns}.
}
