% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{.violin_plot}
\alias{.violin_plot}
\title{Render a violin plot}
\usage{
.violin_plot(x, y)
}
\arguments{
\item{x}{A list with the x-axis values. If it is a nested list, a dropdown-field will be provided in the interactive mode.}

\item{y}{A list with the y-axis values. If it is a nested list, a dropdown-field will be provided in the interactive mode.(Needs to be categorial. Horizontal violinplots are not possible.)}
}
\value{
A list with 1. the plotly object & 2. the data frame used in the plot
}
\description{
Render a violin plot
}
