% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/feature_grid_page.R
\docType{methods}
\name{feature-grid-page}
\alias{feature-grid-page}
\alias{add_feature_grid_page}
\title{Create expression visualization for multiple selected features on a regular grid.}
\usage{
add_feature_grid_page(dashboard, object, ...)

\S4method{add_feature_grid_page}{i2dashboard,missing}(dashboard,
  use_dimred, exprs_values, page = "feature_grid_page",
  title = "Feature grid", menu = "Tools")


  \S4method{add_feature_grid_page}{i2dashboard,SingleCellExperiment}(dashboard,
  object, use_dimred, exprs_values, subset_row = NULL, ...)

\S4method{add_feature_grid_page}{i2dashboard,Seurat}(dashboard, object,
  use_dimred, assay, assay_slot = "data", subset_row = NULL, ...)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{A \linkS4class{SingleCellExperiment::SingleCellExperiment} object or a \linkS4class{Seurat::Seurat} object.}

\item{use_dimred}{A list of data.frames (matrices) or a single data.frame (matrix) containing coordinates of the reduced dimensions, a character vector representing valid \code{reducedDim} slots of \code{object} or names of the \linkS4class{Seurat::DimReduc} object in \code{object@reductions}.}

\item{exprs_values}{A data.frame (matrix) containing expression data of features of interest in rows and samples in columns, or a string representing the name of an \code{assay} of \code{object}.}

\item{page}{The name of the page to be added.}

\item{title}{The title of the page.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}

\item{subset_row}{A character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use. The default of NULL will use all features.}

\item{assay}{A character vector specifying which assay from \code{object@assays} to obtain expression values from.}

\item{slot}{A character vector specifying the name of the slot in the assay.}
}
\value{
An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
Users can select features and a dimension reduction to plot feature expression values.
}
