% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/feature_selection_page.R
\name{feature-selection-page}
\alias{feature-selection-page}
\alias{add_feature_selection_page}
\alias{add_feature_selection_page,i2dashboard,Seurat-method}
\title{Characterize and visualize dimension reductions and sample groupings / metadata.}
\usage{
add_feature_selection_page(dashboard, object, ...)

\S4method{add_feature_selection_page}{i2dashboard,missing}(
  dashboard,
  exprs_values,
  use_function = c("modelGeneVar", "modelGeneVarByPoisson", "modelGeneCV2",
    "modelGeneCV2WithSpikes", "modelGeneVarWithSpikes"),
  title = "Feature selection",
  menu = "Tools",
  ...
)

\S4method{add_feature_selection_page}{i2dashboard,SingleCellExperiment}(dashboard, object, exprs_values = "logcounts", ...)

\S4method{add_feature_selection_page}{i2dashboard,Seurat}(
  dashboard,
  object,
  assay = "RNA",
  slot = "logcounts",
  ...
)
}
\arguments{
\item{dashboard}{A \linkS4class{i2dash::i2dashboard}.}

\item{object}{An object of class \linkS4class{Seurat::Seurat} or \linkS4class{SingleCellExperiment::SingleCellExperiment}.}

\item{...}{Further parameters provided to the method specified in \code{use_function}.}

\item{exprs_values}{A numeric matrix of log-counts, or a string indicating the assay of \code{object}.}

\item{use_function}{Choose the \code{scran} method to model the variance of the log-expression profiles for each gene.}

\item{title}{The title of the page.}

\item{menu}{The name of the menu, under which the page should appear.}

\item{assay}{In case of a \code{Seurat}-class \code{object} a string indicating the assay to use.}

\item{slot}{In case of a \code{Seurat}-class \code{object} a string indicating the assay to use.}
}
\value{
An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
Creates a page with up to four different linked components, including a scatterplot for dimension reductions, a bar plot showing numbers of observations by group, and a silhouette plot to assess grouping consistency.
Additional sample metadata is visualized using boxplots and barplots, depending on the data type of the underlying variable.
}
