% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_by_factors.R
\name{features_by_factors}
\alias{features_by_factors}
\title{Renders a features by factor violin plot}
\usage{
features_by_factors(object, x, y, title = "Features by factor")
}
\arguments{
\item{x}{A list with the x-axis values. If it is a nested list, a dropdown-field will be provided in the interactive mode.}

\item{y}{A list with the y-axis values. If it is a nested list, a dropdown-field will be provided in the interactive mode.(Needs to be categorial. Horizontal violinplots are not possible.)}

\item{plot_title}{The title of the Component}
}
\value{
A string containing markdown code for the rendered textbox
}
\description{
Renders a features by factor violin plot
}
