% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneView_wilson.R
\name{geneview_wilson}
\alias{geneview_wilson}
\title{Method for geneView creation from 'wilson' package}
\usage{
geneview_wilson(object, countTable, group_by, title = NULL, ...)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{countTable}{A matrix with features as rows and observations as columns. The rownames and columnnames should be provided and are used in buiding the heatmap.}

\item{group_by}{A vector with values or a named list will be mapped to the y-axis. In case of a named list, a dropdown menu will be provided in the interactive mode. Note: The length of vectors x and y should be the same as well as the length of all vectors in case of a named list.}

\item{title}{(Optional) The title of the components junk.}

\item{(...)}{Further parameters which are compatible with wilsons create_scatterplot() method. See \code{\link[wilson::create_scatterplot()]{wilson}}.}
}
\value{
A string containing markdown code for the rendered textbox
}
\description{
Method for geneView creation from 'wilson' package
}
