% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\docType{methods}
\name{heatmap}
\alias{heatmap}
\alias{heatmap,i2dashboard,missing-method}
\alias{heatmap,i2dashboard,SingleCellExperiment-method}
\title{Renders a component containing a heatmap based on \link[ComplexHeatmap]{Heatmap}.}
\usage{
heatmap(dashboard, object, ...)

\S4method{heatmap}{i2dashboard,missing}(dashboard, exprs_values,
  column_split = NULL, visualisation_mode = "splitted", title = NULL,
  legend_title = NULL, cluster_rows = FALSE, cluster_columns = FALSE,
  clustering_distance = "euclidean", clustering_method = "complete")

\S4method{heatmap}{i2dashboard,SingleCellExperiment}(dashboard, object,
  exprs_values = "counts", features, column_split = NULL, ...)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{A valid \linkS4class{SingleCellExperiment::SingleCellExperiment} object.}

\item{exprs_values}{A data.frame (matrix) containing expression data of features of interest in rows and samples in columns or a character indicating which assay of the \linkS4class{SingleCellExperiment::SingleCellExperiment} object to use.}

\item{column_split}{An optional data.frame (matrix) with columns containing grouping factors for spliting the columns or a character vector containing the column names of \code{colData(object)} to use. Enables a second visualisation option: heatmap with levels of \code{column_split} as columns and features of \code{exprs_values}.}

\item{visualisation_mode}{Used in case of the static mode and if \code{column_split} is provided: Select between "splitted" or "summarized" heatmap visualisation according to the levels in the 1. column of \code{column_split}.}

\item{title}{Title of the component.}

\item{cluster_rows}{A logical controls whether to make cluster on rows.}

\item{cluster_columns}{A logical controls whether to make cluster on columns.}

\item{clustering_distance}{A pre-defined character which is in ("euclidean", "maximum", "manhattan", "binary", "minkowski").}

\item{clustering_method}{Method to perform hierarchical clustering, pass to \link[stats]{hclust} ("average", "ward.D", "ward.D2", "single", "complete", "mcquitty", "median","centroid").}

\item{features}{A vector with features of interest identical to featurenames in \code{exprs_values} for subsetting \code{exprs_values}.}
}
\value{
A string containing markdown code for the rendered component

An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
Renders a component containing a heatmap based on \link[ComplexHeatmap]{Heatmap}.
}
