% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_wilson.R
\name{heatmap_wilson}
\alias{heatmap_wilson}
\title{Renders a heatmap plot from 'wilson' package}
\usage{
heatmap_wilson(object, countTable, group_by, title = NULL, ...)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{countTable}{A matrix with features as rows and observations as columns. The rownames and columnnames should be provided and are used in buiding the heatmap.}

\item{group_by}{A vector with numerical values or a named list will be mapped to the y-axis. In case of a named list, a dropdown menu will be provided in the interactive mode. Note: The length of vectors x and y should be the same as well as the length of all vectors in case of a named list.}

\item{title}{(Optional) The title of the components junk.}

\item{(...)}{Further parameters which are compatible with wilsons create_scatterplot() method. See \code{\link[wilson::create_scatterplot()]{wilson}}.}
}
\value{
A string containing markdown code for the rendered textbox
}
\description{
Renders a heatmap plot from 'wilson' package
}
