% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotExpression.R
\name{plotExpression}
\alias{plotExpression}
\alias{plotExpression,i2dashboard,missing-method}
\alias{plotExpression,i2dashboard,SingleCellExperiment-method}
\alias{plotExpression,i2dashboard,Seurat-method}
\title{Renders a component containing a plot of the expression values for a set of features (e.g. genes or transcripts), against a continuous or categorical covariate for all cells.(This method is a wrapepr for the function \code{scater::plotExpression})}
\usage{
plotExpression(dashboard, object, ...)

\S4method{plotExpression}{i2dashboard,missing}(
  dashboard,
  exprs_values,
  features = NULL,
  metadata = NULL,
  x = NULL,
  ...
)

\S4method{plotExpression}{i2dashboard,SingleCellExperiment}(
  dashboard,
  object,
  exprs_values = "logcounts",
  features,
  metadata = NULL,
  x = NULL,
  as_plotly = FALSE,
  title = NULL,
  x_title = NULL,
  y_title = NULL,
  plot_title = NULL,
  ncol = 2,
  scales = "fixed",
  ...
)

\S4method{plotExpression}{i2dashboard,Seurat}(dashboard, object, ...)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{An object of class \linkS4class{Seurat::Seurat} or \linkS4class{SingleCellExperiment::SingleCellExperiment}.}

\item{...}{Parameter for \code{scater::plotExpression}.}

\item{exprs_values}{A data.frame (matrix) containing expression data of features of interest in rows and samples in columns, or a string representing the name of an \code{assay} of \code{object}.}

\item{features}{A character vector specifying the features to plot.}

\item{metadata}{A single characters or vector of characters specifying the metadata that is used to group, colour, size, shape the observations.}

\item{x}{A single character specifying a feature or column name from "\code{metadata}" to show on the x-axis.}

\item{as_plotly}{Logical whether the ggplot2 plot should be converted into a plotly plot.}

\item{title}{The title of the component.}

\item{x_title}{The title of the x-axis.}

\item{y_title}{The title of the y-axis.}

\item{plot_title}{The title of the plot.}

\item{ncol}{Integer scalar, specifying the number of columns to be used for the panels of a multi-facet plot.}

\item{scales}{String indicating whether should multi-facet scales be fixed ("\code{fixed}"), free ("\code{free}"), or free in one dimension ("\code{free_x}", "\code{free_y}").}
}
\value{
A string containing markdown code for the rendered component

An object of class \linkS4class{i2dash::i2dashboard}.

A string containing markdown code for the rendered component
}
\description{
Renders a component containing a plot of the expression values for a set of features (e.g. genes or transcripts), against a continuous or categorical covariate for all cells.(This method is a wrapepr for the function \code{scater::plotExpression})
}
\details{
For further information see \code{?scater::plotColData()} / \code{?scater::plotRowData()}
}
