% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plotly_boxplot}
\alias{plotly_boxplot}
\title{Render a box plot with plotly.}
\usage{
plotly_boxplot(df, ..., title_x = NULL, title_group_by = NULL)
}
\arguments{
\item{df}{A dataframe containing the data fo the boxplot.}

\item{...}{these arguments are of either the form value or tag = value and should be valid for the 'plotly::plot_ly()'}

\item{title_x}{A title that describes the observations.}

\item{title_group_by}{A title that describes the grouping factor.}
}
\value{
An object of class \code{plotly}.
}
\description{
Render a box plot with plotly.
}
