% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plotly_scatterplot}
\alias{plotly_scatterplot}
\title{Creates a plotly scatterplot}
\usage{
plotly_scatterplot(..., y_title = NULL, x_title = NULL)
}
\arguments{
\item{...}{these arguments are of either the form value or tag = value and should be valid for the 'plotly::plot_ly()' method.}

\item{y_title}{The title of the x-axis.}

\item{x_title}{The title of the y-axis.}
}
\value{
An object of class \code{plotly}.
}
\description{
Creates a plotly scatterplot
}
