% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plotly_violinplot}
\alias{plotly_violinplot}
\title{Render a vertical violin plot with plotly.}
\usage{
plotly_violinplot(df, group_by = NULL, title_y = NULL,
  title_group_by = NULL)
}
\arguments{
\item{df}{A dataframe containing the data for the violinplot.}

\item{group_by}{A factor, by which observations can be grouped. In case of a named list, a dropdown menu will be provided in the interactive mode.}

\item{title_y}{A title of the y-axis that describes the observations. In case of a named list this parameter is not needed because the names of the list will be used as title of the y axis.}

\item{title_group_by}{A title of the x-axis that describes the grouping factor. In case of a named list this parameter is not needed because the names of the list will be used as title of the x axis.}
}
\value{
An object of class \code{plotly}.
}
\description{
Render a vertical violin plot with plotly.
}
