% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMetadata.R
\name{plotMetadata}
\alias{plotMetadata}
\alias{plotMetadata,i2dashboard,missing-method}
\alias{plotMetadata,i2dashboard,SingleCellExperiment-method}
\title{Renders a component containing a violinplot with column- or row-level metadata.(This method is a wrapepr for the function \code{scater::plotColData} or \code{scater::plotRowData})}
\usage{
plotMetadata(dashboard, object, ...)

\S4method{plotMetadata}{i2dashboard,missing}(
  dashboard,
  y,
  x = NULL,
  from = c("colData", "rowData"),
  metadata = NULL,
  ...
)

\S4method{plotMetadata}{i2dashboard,SingleCellExperiment}(
  dashboard,
  object,
  from = c("colData", "rowData"),
  y,
  x = NULL,
  metadata = NULL,
  group_by = NULL,
  as_plotly = FALSE,
  title = NULL,
  x_title = NULL,
  y_title = NULL,
  plot_title = NULL
)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{An object of class \linkS4class{Seurat::Seurat} or \linkS4class{SingleCellExperiment::SingleCellExperiment}.}

\item{y}{A single characters or vector of characters specifying the metadata to show on the y-axis.}

\item{x}{A single characters or vector of characters specifying the metadata to show on the x-axis.}

\item{from}{A character specifying whether the function \code{scater::plotColData} or \code{scater::plotRowData} is used.}

\item{metadata}{A single characters or vector of characters specifying the metadata that is used to colour, size, shape the observations.}

\item{group_by}{A single character specifying the metadata from \code{metadata} that is used to to group the observations.}

\item{x_title}{The title of the x-axis.}

\item{y_title}{The title of the y-axis.}

\item{plot_title}{The title of the component.}
}
\value{
A string containing markdown code for the rendered component

An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
Renders a component containing a violinplot with column- or row-level metadata.(This method is a wrapepr for the function \code{scater::plotColData} or \code{scater::plotRowData})
}
\details{
For further information see \code{?scater::plotColData()} / \code{?scater::plotRowData()}
}
