% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\alias{summarize_samples}
\alias{summarize_features}
\title{Summarize samples or features of objects containing data from single-cell experiments.}
\usage{
summarize_samples(object, ...)

summarize_features(object, ...)
}
\arguments{
\item{object}{A SingleCellExperiment or Seurat object.}

\item{columns}{The metadata columns to summarize.}

\item{FUNS}{A named vector indicating summary functions.}

\item{group_by}{(Optional) A vector, which is used for grouping.}
}
\value{
A \code{kableExtra} object with one row for each \code{columns} and one column for each \code{FUNS}, containing the desired summary. Optionaly \code{group_by} can be provided. In this case for each level in \code{group_by} a column is created with the summerized values for each \code{FUNS}.
}
\description{
Summarize samples or features of objects containing data from single-cell experiments.
}
