% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_data_}
\alias{summarize_data_}
\title{Summarizes columns of \code{data} using \code{FUNS}.}
\usage{
summarize_data_(data, columns, FUNS = c(mean = "mean", median =
  "median"))
}
\arguments{
\item{data}{A data.frame.}

\item{columns}{The columns of \code{data} that are summarized.}

\item{FUNS}{A named vector of summary functions}
}
\value{
A \code{data.frame} that contains summary statistics.
}
\description{
Summarizes columns of \code{data} using \code{FUNS}.
}
