% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/summarize_discarded.R
\name{summarize_discarded}
\alias{summarize_discarded}
\alias{summarize_discarded,i2dashboard,missing-method}
\alias{summarize_discarded,i2dashboard,SingleCellExperiment-method}
\alias{summarize_discarded,i2dashboard,Seurat-method}
\title{Renders a componnet containing an HTML table of summarized sample or feature metadata of single-cell experiment objects or data.frames.}
\usage{
summarize_discarded(dashboard, object, ...)

\S4method{summarize_discarded}{i2dashboard,missing}(
  dashboard,
  data,
  columns,
  discarded,
  footnote_title = NULL,
  footnote_text = NULL,
  table_title = "Number of kept / discarded cells per condition.",
  colnames = c("condition", "kept", "discarded"),
  title = NULL
)

\S4method{summarize_discarded}{i2dashboard,SingleCellExperiment}(
  dashboard,
  object,
  discarded,
  from = c("colData", "rowData"),
  columns = NULL,
  ...
)

\S4method{summarize_discarded}{i2dashboard,Seurat}(
  dashboard,
  object,
  discarded,
  from = c("colData", "rowData"),
  columns = NULL,
  assay = "RNA",
  ...
)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{An object of class \linkS4class{Seurat::Seurat} or \linkS4class{SingleCellExperiment::SingleCellExperiment}.}

\item{data}{Data.frame containing numeric observations for summary.}

\item{columns}{The columns of \code{data} containing experiemntal factors.}

\item{discarded}{A column of \code{data} contatining logical values.}

\item{footnote_title}{Title of the footnote.}

\item{footnote_text}{Description for the table shown as footnote.}

\item{table_title}{Title of the table.}

\item{title}{The title of the component.}

\item{assay}{A character defining the assay of \code{object}.}
}
\value{
A string containing markdown code for the rendered component
}
\description{
Renders a componnet containing an HTML table of summarized sample or feature metadata of single-cell experiment objects or data.frames.
}
