% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/summarize.R
\name{summarize_metadata}
\alias{summarize_metadata}
\alias{summarize_metadata,i2dashboard,missing-method}
\alias{summarize_metadata,i2dashboard,SingleCellExperiment-method}
\alias{summarize_metadata,i2dashboard,Seurat-method}
\title{Renders a componnet containing an HTML table of summarized sample or feature metadata of single-cell experiment objects or data.frames.}
\usage{
summarize_metadata(dashboard, object, ...)

\S4method{summarize_metadata}{i2dashboard,missing}(
  dashboard,
  data,
  columns = colnames(data),
  FUNS = c("min", "max", "mean", "median"),
  group_by = NULL,
  footnote_title = NULL,
  footnote_text = NULL,
  title = NULL
)

\S4method{summarize_metadata}{i2dashboard,SingleCellExperiment}(
  dashboard,
  object,
  from = c("colData", "rowData"),
  columns = NULL,
  group_by = NULL,
  footnote_title = NULL,
  footnote_text = NULL,
  ...
)

\S4method{summarize_metadata}{i2dashboard,Seurat}(
  dashboard,
  object,
  from = c("meta.data", "meta.features"),
  assay = "RNA",
  columns = NULL,
  group_by = NULL,
  footnote_title = NULL,
  footnote_text = NULL,
  ...
)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{An object of class \linkS4class{Seurat::Seurat} or \linkS4class{SingleCellExperiment::SingleCellExperiment}.}

\item{data}{Data.frame containing numeric observations for summary.}

\item{columns}{The metadata columns to summarize.}

\item{FUNS}{A named vector indicating summary functions.}

\item{group_by}{Optionally provide data containing grouping factors for grouping.}

\item{footnote_title}{Provide a title for the table description.}

\item{footnote_text}{Provide a string for table description.}

\item{title}{The title of the component.}

\item{assay}{A character defining the assay of \code{object}.}
}
\value{
A string containing markdown code for the rendered component
}
\description{
Renders a componnet containing an HTML table of summarized sample or feature metadata of single-cell experiment objects or data.frames.
}
