% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/verbatim_text.R
\docType{methods}
\name{verbatim_text}
\alias{verbatim_text}
\alias{verbatim_text,i2dashboard,missing-method}
\title{Renders a component containing a shiny verbatim_text.}
\usage{
verbatim_text(dashboard, object, ...)

\S4method{verbatim_text}{i2dashboard,missing}(dashboard,
  transmitter = NULL, event = "plotly_selected", title = NULL)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{transmitter}{A character defining the id of an existing transmitter to obtain the data from.}

\item{event}{The type of plotly event. (see \code{?plotly::event_data()}.}

\item{title}{Title of the component.}
}
\value{
A string containing markdown code for the rendered component
}
\description{
Renders a component containing a shiny verbatim_text.
}
