% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression_page.R
\name{add_gene_expression_page}
\alias{add_gene_expression_page}
\title{Renders a linked gene expression page.}
\usage{
add_gene_expression_page(object, reduced_dim, count_table, metadata,
  grouping, title = NULL, labels = NULL, menu = NULL,
  sidebar = NULL)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{reduced_dim}{A data.frame or matrix containing two columns for the x and y coordinates. The rownames are used as labels.}

\item{count_table}{A data.frame or matrix containing the genes of interest as rownames and cells/ samples as columnnamens.}

\item{metadata}{A data.frame or matrix with factorial cell metadata or a vector with values (e.g. cluster, timepoint, etc.). Should have the same rownumber as 'reduced_dim'.}

\item{grouping}{A character string identical to one of the column names in 'metadata'.}

\item{title}{The title of the page.}

\item{labels}{(Optional) A vector with labels. Should be of the same length as the rownumber of 'reduced_dim'.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}

\item{sidebar}{(Optional) The page layout (see below).}
}
\value{
A string containing markdown code for the rendered page.
}
\description{
Renders a linked gene expression page.
}
