% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/gene_expression_page.R
\docType{methods}
\name{add_gene_expression_page}
\alias{add_gene_expression_page}
\alias{add_gene_expression_page,i2dashboard,missing-method}
\alias{add_gene_expression_page,i2dashboard,SingleCellExperiment-method}
\alias{add_gene_expression_page,i2dashboard,Seurat-method}
\title{Add a gene expression page.}
\usage{
add_gene_expression_page(dashboard, object, ...)

\S4method{add_gene_expression_page}{i2dashboard,missing}(dashboard,
  use_dimred, exprs_values, group_by = NULL,
  labels = rownames(use_dimred), title = "Gene expression",
  menu = NULL)


  \S4method{add_gene_expression_page}{i2dashboard,SingleCellExperiment}(dashboard,
  object, use_dimred, exprs_values, metadata_columns = NULL,
  features = NULL, title = "Gene expression", menu = NULL)

\S4method{add_gene_expression_page}{i2dashboard,Seurat}(dashboard, object,
  use_dimred, assay, metadata_columns, assay_slot = "data",
  subset_row = NULL, title = NULL, menu = NULL)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{A valid \linkS4class{SingleCellExperiment::SingleCellExperiment} object.}

\item{use_dimred}{A data.frame (matrix) containing coordinates of the reduced dimensions.}

\item{exprs_values}{A data.frame (matrix) containing expression data of features of interest in rows and samples in columns.}

\item{group_by}{A data.frame (matrix) along rows of \code{use_dimred} that is used for grouping expression values in the violin plot.}

\item{labels}{A vector with optional sample labels that are used instead of \code{rownames(use_dimred)}.}

\item{title}{The title of the page.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}

\item{metadata_columns}{A character vector with column names of \code{colData(object)} to use for cell grouping.}

\item{features}{A character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use. The default of NULL will use all features.}

\item{assay}{A character vector specifying which assay from \code{object@assays} to obtain expression values from.}

\item{subset_row}{A character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use. The default of NULL will use all features.}

\item{dimred}{A string or integer scalar indicating the reduced dimension result in \code{reducedDims(object)}.}

\item{exprs_values}{A string or integer scalar specifying which assay to obtain expression values from.}

\item{object}{A valid \linkS4class{Seurat::Seurat} object.}

\item{dimred}{A character vector indicating the \linkS4class{Seurat::DimReduc} object in \code{object@reductions}.}

\item{metadata_columns}{A character vector with column names of \code{object@meta.data} to use for cell grouping.}

\item{slot}{A character vector specifying the name of the slot in the assay.}
}
\value{
An object of class \linkS4class{i2dash::i2dashboard}.

An object of class \linkS4class{i2dash::i2dashboard}.

An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
This function adds a page with two linked components to the \code{dashboard} object: A scatterplot, showing samples in along two-dimensional coordinates, and a violin plot, showing feature expression values by groups defined in \code{group_by}.
}
