% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot}
\alias{barplot}
\title{Renders a horizontal barplot}
\usage{
barplot(object, group_by, x = NULL, title = NULL, title_x = NULL,
  title_group_by = NULL)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{group_by}{A vector with factorial values or characters or a named list, which will be used for grouping the observations. In case of a named list, a dropdown menu will be provided in the interactive mode. Note: If parameter 'x' is not provided (NULL), a barplot with the number of each 'level' is used as observation.}

\item{x}{(Optional) A vector with numerical values or a named list, which represents the observations for the horizontal barplot (x-axis). In case of a named list, a dropdown menu will be provided in the interactive mode.}

\item{title}{(Optional) The title of the components junk.}

\item{title_x}{(Optional) The title of the x-axis.}

\item{title_group_by}{(Optional) The title of the y-axis.}
}
\value{
A string containing markdown code for the rendered component.
}
\description{
Renders a horizontal barplot
}
