% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot}
\alias{barplot}
\title{Renders a barplot}
\usage{
barplot(object, group_by, x = NULL, title = NULL,
  title_group_by = NULL, title_x = NULL)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{group_by}{A factor, by which observations are grouped. In case of a named list, a dropdown menu will be provided in the interactive mode.}

\item{x}{Numeric observations for the boxplot. In case of a named list, a dropdown menu will be provided in the interactive mode.}

\item{title}{The title of the components junk.}
}
\value{
A string containing markdown code for the rendered textbox
}
\description{
Renders a barplot
}
