% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/barplot.R
\docType{methods}
\name{barplot}
\alias{barplot}
\alias{barplot,i2dashboard,missing-method}
\alias{barplot,i2dashboard,SingleCellExperiment-method}
\alias{barplot,i2dashboard,Seurat-method}
\title{Renders a component containing a horizontal barplot.}
\usage{
barplot(dashboard, object, ...)

\S4method{barplot}{i2dashboard,missing}(dashboard, y_group_by,
  x_group_by = NULL, title = NULL, x_group_by_title = NULL,
  y_group_by_title = NULL, transmitter = NULL)

\S4method{barplot}{i2dashboard,SingleCellExperiment}(dashboard, object,
  use = c("colData", "rowData"), y_group_by = NULL,
  x_group_by = NULL, ...)

\S4method{barplot}{i2dashboard,Seurat}(dashboard, object,
  use = c("meta.data", "meta.features"), assay = "RNA",
  y_group_by = NULL, x_group_by = NULL, ...)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{An object of class \linkS4class{Seurat::Seurat} or \linkS4class{SingleCellExperiment::SingleCellExperiment}.}

\item{y_group_by}{A data.frame (matrix) with columns containing grouping factors for the vertical axis or a character vector indicating the columns to use from \code{use}.}

\item{x_group_by}{Optionally provide a data.frame (matrix) with columns containing grouping factors for the horizontal axis or a character vector indicating the columns to use from \code{use}. The result is a barplot grouped by the levels in \code{x_group_by}, which shows the relative number of its observations.}

\item{title}{The title of the components junk.}

\item{x_group_by_title}{The title of the x-axis.}

\item{y_group_by_title}{The title of the y-axis.}

\item{transmitter}{A character defining the id of an existing transmitter to obtain the data from.}

\item{use}{A character specifying where to obtain the data from. Valid input for SingleCellExperiment object: ("colData", "rowData"). Valid input for Seurat object: ("meta.data" for sample metadata, "meta.feature" for feature metadata.)}

\item{assay}{Necessery, if \code{use} = "meta.feature". A character defining the assay to obtain the feature metadata from (default "RNA").}
}
\value{
A string containing markdown code for the rendered component

An object of class \linkS4class{i2dash::i2dashboard}.

An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
Renders a component containing a horizontal barplot.
}
