% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/barplot.R
\name{barplot}
\alias{barplot}
\alias{barplot,i2dashboard,missing-method}
\alias{barplot,i2dashboard,SingleCellExperiment-method}
\alias{barplot,i2dashboard,Seurat-method}
\title{Renders a component containing a horizontal barplot.}
\usage{
barplot(dashboard, object, ...)

\S4method{barplot}{i2dashboard,missing}(
  dashboard,
  y_group_by,
  x_group_by = NULL,
  title = NULL,
  x_group_by_title = NULL,
  y_group_by_title = NULL
)

\S4method{barplot}{i2dashboard,SingleCellExperiment}(
  dashboard,
  object,
  y_group_by = NULL,
  x_group_by = NULL,
  from = c("colData", "rowData"),
  ...
)

\S4method{barplot}{i2dashboard,Seurat}(
  dashboard,
  object,
  y_group_by = NULL,
  x_group_by = NULL,
  from = c("meta.data", "meta.features"),
  assay = "RNA",
  ...
)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dashboard}.}

\item{object}{An object of class \linkS4class{Seurat} or \linkS4class{SingleCellExperiment}.}

\item{...}{Further parameters passed to the core function.}

\item{y_group_by}{Data containing grouping factors for the vertical axis.}

\item{x_group_by}{Optionally provide data containing grouping factors for the horizontal axis. The result is a barplot grouped by the levels in \code{x_group_by} and shows the relative number of its observations.}

\item{title}{The title of the components junk.}

\item{x_group_by_title}{The title of the x-axis.}

\item{y_group_by_title}{The title of the y-axis.}

\item{from}{A character specifying where to obtain the data from (see Details).}

\item{assay}{A character defining the assay of \code{object} and is used for obtaining the meta data of the respective assay (default "RNA") (see Details).}
}
\value{
A string containing markdown code for the rendered component
}
\description{
Renders a component containing a horizontal barplot.
}
\details{
The parameters \code{y_group_by}, \code{x_group_by} take different arguments depending on the class of \code{object}.
  In case of the \emph{i2dashboard,missing}-method, the parameters are expected to be of class \code{data.frame} or \code{matrix}.
  In case of the \emph{i2dashboard,SingleCellExperiment} or \emph{i2dashboard,Seurat} method, the parameters are expected to be of class \code{character} and
  \itemize{
    \item \code{from} can either be \code{"colData"} or \code{"rowData"} for a \linkS4class{SingleCellExperiment} object or
    \item \code{"meta.data"} or \code{"meta.features"} for a \linkS4class{Seurat} object.
  }
  In both cases, \code{y_group_by} and \code{x_group_by} take column names of \code{from}.
}
