% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/barplot.R
\name{barplot}
\alias{barplot}
\alias{barplot,i2dashboard,missing-method}
\alias{barplot,i2dashboard,SingleCellExperiment-method}
\title{Renders a component containing a horizontal barplot.}
\usage{
barplot(dashboard, object, ...)

\S4method{barplot}{i2dashboard,missing}(
  dashboard,
  y_group_by,
  x_group_by = NULL,
  title = NULL,
  x_group_by_title = NULL,
  y_group_by_title = NULL
)

\S4method{barplot}{i2dashboard,SingleCellExperiment}(
  dashboard,
  object,
  use = "colData",
  y_group_by = NULL,
  x_group_by = NULL,
  ...
)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{A valid \linkS4class{SingleCellExperiment::SingleCellExperiment} object.}

\item{y_group_by}{A data.frame (matrix) with columns containing grouping factors for the vertical axis.}

\item{x_group_by}{Optionally provide a data.frame (matrix) with columns containing grouping factors for the horizontal axis. The result is a barplot grouped by the levels in \code{x_group_by} and shows the relative number of its observations.}

\item{title}{The title of the components junk.}

\item{x_group_by_title}{The title of the x-axis.}

\item{y_group_by_title}{The title of the y-axis.}

\item{use}{A character specifying where to obtain the data from. One of \code{"colData"} or \code{"rowData"}.}
}
\value{
A string containing markdown code for the rendered component

An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
Renders a component containing a horizontal barplot.
}
