% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{Renders a boxplot}
\usage{
boxplot(object, x, group_by = NULL, title = NULL, title_x = NULL,
  title_group_by = NULL)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{x}{Numeric observations for the boxplot. In case of a named list, a dropdown menu will be provided in the interactive mode.}

\item{group_by}{A factor, by which observations can optionally be grouped. In case of a named list, a dropdown menu will be provided in the interactive mode.}

\item{title}{The title of the components junk.}
}
\value{
A string containing markdown code for the rendered component.
}
\description{
Renders a boxplot
}
