% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/boxplot.R
\name{boxplot}
\alias{boxplot}
\alias{boxplot,i2dashboard,missing-method}
\alias{boxplot,i2dashboard,SingleCellExperiment-method}
\title{Renders a component containing a boxplot}
\usage{
boxplot(dashboard, object, ...)

\S4method{boxplot}{i2dashboard,missing}(
  dashboard,
  x,
  group_by = NULL,
  title = NULL,
  x_title = NULL,
  group_by_title = NULL
)

\S4method{boxplot}{i2dashboard,SingleCellExperiment}(
  dashboard,
  object,
  use = "colData",
  x = NULL,
  group_by = NULL,
  title = NULL,
  x_title = NULL,
  group_by_title = NULL
)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{A valid \linkS4class{SingleCellExperiment::SingleCellExperiment} object.}

\item{x}{A data.frame (matrix) containing numeric observations for the horizontal axis, or a character vector indicating column names of \code{colData(object)}, \code{rowData(object)}.}

\item{group_by}{An optional data.frame (matrix) with columns containing grouping factors for the vertical axis.}

\item{title}{The title of the components junk.}

\item{x_title}{The title of the x-axis.}

\item{group_by_title}{The title of the y-axis.}

\item{use}{A character specifying where to obtain the data from. One of \code{"colData"} or \code{"rowData"}.}
}
\value{
A string containing markdown code for the rendered component

An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
Renders a component containing a boxplot
}
