% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot.R
\name{bubbleplot}
\alias{bubbleplot}
\title{Renders a component containing a bubbleplot with optional selection options}
\usage{
bubbleplot(
  dashboard,
  x,
  y,
  size,
  colour_by = NULL,
  labels = NULL,
  title = NULL,
  x_title = NULL,
  y_title = NULL
)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{x}{A data.frame (matrix) containing columns with numeric values that will be mapped to the x-axis.}

\item{y}{A data.frame (matrix) containing columns with numeric values that will be mapped to the y-axis.}

\item{size}{A ata.frame (matrix) containing columns with numeric values that describe the size of the bubbles.}

\item{colour_by}{An optional data.frame (matrix) containing columns with numeric or factorial values that will be used for colouring.}

\item{labels}{An optional vector with sample names.}

\item{title}{The title of the components junk.}

\item{x_title}{An optional title of the x-axis. If not provided the column names from \code{x} are used instead.}

\item{y_title}{An optional title of the y-axis. If not provided the column names from \code{y}  are used instead.}
}
\value{
A string containing markdown code for the rendered component
}
\description{
Renders a component containing a bubbleplot with optional selection options
}
