% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dimred_comparison_page.R
\name{add_dimred_comparison_page}
\alias{add_dimred_comparison_page}
\alias{add_dimred_comparison_page,i2dashboard,missing-method}
\alias{add_dimred_comparison_page,i2dashboard,SingleCellExperiment-method}
\alias{add_dimred_comparison_page,i2dashboard,SummarizedExperiment-method}
\alias{add_dimred_comparison_page,i2dashboard,Seurat-method}
\title{Explore the effects of the parameters "theta" and "perplexity" of a t-stochastic neighbour embedding or "n_neighors" of a UMAP embedding.}
\usage{
add_dimred_comparison_page(dashboard, object, ...)

\S4method{add_dimred_comparison_page}{i2dashboard,missing}(
  dashboard,
  exprs_values,
  seed = NULL,
  calculateUMAP = list(),
  calculateTSNE = list(),
  page = "dimred_comparison_page",
  title = "Explore UMAP/t-SNE parameter",
  menu = "Tools"
)

\S4method{add_dimred_comparison_page}{i2dashboard,SingleCellExperiment}(
  dashboard,
  object,
  exprs_values = "logcounts",
  subset_row = NULL,
  ...
)

\S4method{add_dimred_comparison_page}{i2dashboard,SummarizedExperiment}(
  dashboard,
  object,
  exprs_values = "logcounts",
  subset_row = NULL,
  ...
)

\S4method{add_dimred_comparison_page}{i2dashboard,Seurat}(
  dashboard,
  object,
  assay = "RNA",
  assay_slot = "data",
  subset_row = NULL,
  ...
)
}
\arguments{
\item{dashboard}{A \linkS4class{i2dashboard}.}

\item{object}{An object of class \linkS4class{Seurat}, \linkS4class{SingleCellExperiment} or \linkS4class{SummarizedExperiment}.}

\item{...}{Further parameters passed to the core function.}

\item{exprs_values}{A numeric matrix of log-expression values where rows are features and columns are cells. Alternatively, a character indicating which assay of \code{object} to use.}

\item{seed}{An integer vector, containing the random number generator (RNG) state for random number generation with \code{set.seed()}.}

\item{calculateUMAP}{A list with parameters passed on to \code{scater::calculateUMAP()}}

\item{calculateTSNE}{A list with parameters passed on to \code{scater::calculateTSNE()}}

\item{page}{The name of the page to be added.}

\item{title}{The title of the page.}

\item{menu}{The menu tab to which this page is subordinated. Default "Tools"}

\item{subset_row}{An optional character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use. The default of NULL will use all features.}

\item{assay}{A character specifying the assay (\code{object@assays}) to obtain expression values from. (Default: "RNA")}

\item{assay_slot}{A character specifying the name of the data slot in the assay. (Default: "data")}
}
\value{
An object of class \linkS4class{i2dashboard}.
}
\description{
Creates a page with two tabs and a sidebar. The first tab generates the dimension reduction plot of the UMAP or t-SNE with the parameters set in the sidebar by clicking the "Generate plot" button. It is possible to save the scatterplot with its parameters to compare it with another parameters by clicking the button "Add plot for comparison". The second tab "Compare selected plots" contains a grid layout with the saved reduced dimensions. A list in the sidebar shows the saved plots and enables the deletion of plots.
}
