% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dimension_reduction_page.R,
%   R/dimred_metadata_page.R
\name{dimred-metadata-page}
\alias{dimred-metadata-page}
\alias{add_dimred_metadata_page}
\alias{add_dimred_feature_page}
\title{Add a dimension reduction page with feature metadata}
\usage{
add_dimred_feature_page(dashboard, object, ...)

add_dimred_metadata_page(dashboard, object, ...)

\S4method{add_dimred_metadata_page}{i2dashboard,missing}(
  dashboard,
  use_dimred,
  metadata,
  group_by,
  title = "Dimension reduction",
  labels = NULL,
  show_group_sizes = TRUE,
  show_silhouette = FALSE,
  menu = NULL
)

\S4method{add_dimred_feature_page}{i2dashboard,missing}(
  dashboard,
  use_dimred,
  exprs_values,
  feature_metadata,
  title = "Marker gene expression",
  menu = NULL
)

\S4method{add_dimred_feature_page}{i2dashboard,SingleCellExperiment}(
  dashboard,
  object,
  use_dimred,
  exprs_values,
  feature_metadata,
  subset_row,
  title = "Marker gene expression",
  menu = NULL
)

\S4method{add_dimred_feature_page}{i2dashboard,Seurat}(
  dashboard,
  object,
  use_dimred,
  exprs_values,
  feature_metadata,
  subset_row,
  assay,
  assay_slot = "data",
  title = "Marker gene expression",
  menu = NULL
)
}
\arguments{
\item{dashboard}{A \linkS4class{i2dash::i2dashboard}.}

\item{object}{A \linkS4class{SingleCellExperiment::SingleCellExperiment} object or a \linkS4class{Seurat::Seurat} object.}

\item{use_dimred}{A data.frame (matrix) containing coordinates of the reduced dimensions. Rownames are used as sample labels.}

\item{metadata}{A data.frame (matrix) containing metadata (e.g. cluster, timepoint, number of features, etc) along samples.}

\item{group_by}{A string indicating a column in \code{metadata} that is used to group observations.}

\item{title}{The title of the page.}

\item{labels}{An optional vector with sample labels.}

\item{show_group_sizes}{A logical value indicating if a barplot showing the number of observations from \code{group_by} will be creaed (default \code{TRUE}).}

\item{show_silhouette}{A logical value indicating if a silhouette plot should be shown (default \code{FALSE}).}

\item{menu}{The name of the menu, under which the page should appear.}

\item{exprs_values}{A data.frame (matrix) containing expression data of features of interest in rows and samples in columns, or a string representing the name of an \code{assay} of \code{object}.}

\item{feature_metadata}{A data.frame (matrix) along rows of \code{exprs_values} containing feature metadata, or a character vector indicating columns from \code{rowData(object)} or \code{object@meta.data}.}

\item{assay}{A character vector specifying which assay from \code{object@assays} to obtain expression values from.}

\item{slot}{A character vector specifying the name of the slot in the assay.}
}
\value{
An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
This function adds a page with two linked components to the \code{dashboard} object: A scatterplot, showing samples in along two-dimensional coordinates, and a table, showing feature metadata. A click on a feature in the table updates the scatterplot with the feature expression.

Creates a page with up to four different linked components, including a scatterplot for dimension reductions, a bar plot showing numbers of observations by group, and a silhouette plot to assess grouping consistency.
Additional sample metadata is visualized using boxplots and barplots, depending on the data type of the underlying variable.
}
