% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dimred_metadata_page.R
\docType{methods}
\name{dimred-metadata-page}
\alias{dimred-metadata-page}
\alias{add_dimred_metadata_page}
\title{View a dimension reduction side-by-side with feature metadata}
\usage{
add_dimred_metadata_page(dashboard, object, ...)

\S4method{add_dimred_metadata_page}{i2dashboard,missing}(dashboard,
  use_dimred, exprs_values, feature_metadata,
  title = "Marker gene expression", menu = NULL)


  \S4method{add_dimred_metadata_page}{i2dashboard,SingleCellExperiment}(dashboard,
  object, use_dimred, exprs_values, feature_metadata, subset_row,
  title = "Marker gene expression", menu = NULL)

\S4method{add_dimred_metadata_page}{i2dashboard,Seurat}(dashboard, object,
  use_dimred, exprs_values, feature_metadata, subset_row, assay,
  assay_slot = "data", title = "Marker gene expression", menu = NULL)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{A \linkS4class{SingleCellExperiment::SingleCellExperiment} object or a \linkS4class{Seurat::Seurat} object.}

\item{use_dimred}{A list of data.frames (matrices) or a single data.frame (matrix) containing coordinates of the reduced dimensions, a character vector representing valid \code{reducedDim} slots of \code{object} or names of the \linkS4class{Seurat::DimReduc} object in \code{object@reductions}.}

\item{exprs_values}{A data.frame (matrix) containing expression data of features of interest in rows and samples in columns, or a string representing the name of an \code{assay} of \code{object}.}

\item{feature_metadata}{A data.frame (matrix) along rows of \code{exprs_values} containing feature metadata, or a character vector indicating columns from \code{rowData(object)} or \code{object@meta.data}.}

\item{title}{The title of the page.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}

\item{assay}{A character vector specifying which assay from \code{object@assays} to obtain expression values from.}

\item{slot}{A character vector specifying the name of the slot in the assay.}
}
\description{
The dimension reduction plot is colored by feature expression and updated if the users clicks feature rows in the metadata table.
}
