% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dimension_reduction_page.R
\docType{methods}
\name{dimred-metadata-page}
\alias{dimred-metadata-page}
\alias{add_dimred_metadata_page}
\title{Characterize and visualize dimension reductions and sample groupings / metadata.}
\usage{
add_dimred_metadata_page(dashboard, object, ...)

\S4method{add_dimred_metadata_page}{i2dashboard,missing}(dashboard,
  use_dimred, metadata, group_by, title = NULL, labels = NULL,
  show_group_sizes = TRUE, show_silhouette = FALSE, menu = NULL)
}
\arguments{
\item{dashboard}{A \linkS4class{i2dash::i2dashboard}.}

\item{use_dimred}{A data.frame (matrix) containing coordinates of the reduced dimensions. Rownames are used as sample labels.}

\item{metadata}{A data.frame (matrix) containing metadata (e.g. cluster, timepoint, number of features, etc) along samples.}

\item{group_by}{A string indicating a column in \code{metadata} that is used to group observations.}

\item{title}{The title of the page.}

\item{labels}{An optional vector with sample labels.}

\item{show_group_sizes}{A logical value indicating if a barplot showing the number of observations from \code{group_by} will be creaed (default \code{TRUE}).}

\item{show_silhouette}{A logical value indicating if a silhouette plot should be shown (default \code{FALSE}).}

\item{menu}{The name of the menu, under which the page should appear.}
}
\value{
An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
Creates a page with up to four different linked components, including a scatterplot for dimension reductions, a bar plot showing numbers of observations by group, and a silhouette plot to assess grouping consistency.
Additional sample metadata is visualized using boxplots and barplots, depending on the data type of the underlying variable.
}
