% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dimred_sample_page.R
\name{dimred-sample-page}
\alias{dimred-sample-page}
\alias{add_dimred_sample_page}
\title{Characterize and visualize dimension reductions and sample groupings / metadata.}
\usage{
add_dimred_sample_page(dashboard, object, ...)

\S4method{add_dimred_sample_page}{i2dashboard,missing}(
  dashboard,
  use_dimred,
  sample_metadata,
  group_by,
  page = "dimred_sample_page",
  title = "Dim. reduction & sample metadata",
  labels = rownames(use_dimred),
  show_group_sizes = TRUE,
  show_silhouette = FALSE,
  menu = NULL
)

\S4method{add_dimred_sample_page}{i2dashboard,SingleCellExperiment}(dashboard, object, use_dimred, sample_metadata, ...)

\S4method{add_dimred_sample_page}{i2dashboard,Seurat}(dashboard, object, use_dimred, sample_metadata, ...)
}
\arguments{
\item{dashboard}{A \linkS4class{i2dash::i2dashboard}.}

\item{object}{An object of class \linkS4class{Seurat::Seurat} or \linkS4class{SingleCellExperiment::SingleCellExperiment}.}

\item{use_dimred}{Data containing coordinates of the reduced dimensions or a string indicating a dimension reduction from "reductions" of a Seurat \code{object}. Rownames are used as sample labels.}

\item{sample_metadata}{Sample metadata in columns and samples in rows (see Details).}

\item{group_by}{A string indicating a column in \code{sample_metadata} that is used to group observations.}

\item{page}{The name of the page to be added.}

\item{title}{The title of the page.}

\item{labels}{An optional vector with sample labels.}

\item{show_group_sizes}{A logical value indicating if a barplot showing the number of observations from \code{group_by} will be creaed (default \code{TRUE}).}

\item{show_silhouette}{A logical value indicating if a silhouette plot should be shown (default \code{FALSE}).}

\item{menu}{The name of the menu, under which the page should appear.}
}
\value{
An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
Creates a page with up to four different linked components, including a scatterplot for dimension reductions, a bar plot showing numbers of observations by group, and a silhouette plot to assess grouping consistency.
Additional sample metadata is visualized using boxplots and barplots, depending on the data type of the underlying variable.
}
\details{
The parameters \code{use_dimred}, \code{sample_metadata} (or \code{assay}) and \code{group_by} take different arguments depending on the class of \code{object}.
  In case no object is supplied (\emph{i2dashboard,missing}-method), the parameters are expected to be of class \code{data.frame} or \code{matrix}.
  In case a \linkS4class{SingleCellExperiment::SingleCellExperiment} or \linkS4class{Seurat::Seurat} object is supplied, the parameters are expected to be of class \code{character}, containing
  \itemize{
    \item the name of an item in \code{reducedDims(object)} or \code{object@reductions},
    \item a valid assay name from \code{assayNames(object)} or \code{names(object@assays)},
    \item column names of \code{colData(object)} or \code{object@meta.data}.
  }
}
