% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/feauture_expression_page.R
\docType{methods}
\name{feature-expression-page}
\alias{feature-expression-page}
\alias{add_feature_expression_page}
\title{Add a feature expression page.}
\usage{
add_feature_expression_page(dashboard, object, ...)

\S4method{add_feature_expression_page}{i2dashboard,missing}(dashboard,
  use_dimred, exprs_values, page = "feature_expression_page",
  group_by = NULL, labels = rownames(use_dimred),
  title = "Feature expression", menu = NULL)


  \S4method{add_feature_expression_page}{i2dashboard,SingleCellExperiment}(dashboard,
  object, use_dimred, exprs_values, metadata_columns = NULL,
  subset_row = NULL, ...)

\S4method{add_feature_expression_page}{i2dashboard,Seurat}(dashboard,
  object, use_dimred, group_by = NULL, assay = "RNA",
  assay_slot = "data", subset_row = NULL, ...)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{A \linkS4class{SingleCellExperiment::SingleCellExperiment} object or a \linkS4class{Seurat::Seurat} object.}

\item{use_dimred}{A data.frame (matrix) containing coordinates of the reduced dimension or a character representing a valid \code{reducedDim} slot of the SingleCellExperiment \code{object}/ a name of the \linkS4class{Seurat::DimReduc} object in \code{object@reductions}.}

\item{exprs_values}{A data.frame (matrix) containing expression data of features of interest in rows and samples in columns.}

\item{page}{The name of the page to be added.}

\item{group_by}{A data.frame (matrix) containing factorial metadata (e.g. cluster, timepoint, etc.) along samples or a character vector indicating the columns to use from the "meta.data" of a Seurat \code{object}.}

\item{labels}{A vector with optional sample labels that are used instead of \code{rownames(use_dimred)}.}

\item{title}{The title of the page.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}

\item{assay}{A character specifying the assay (\code{object@assays}) to obtain expression values from. (Default: "RNA")}

\item{assay_slot}{A character specifying the name of the data slot in the assay. (Default: "data")}
}
\value{
An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
This function adds a page with two linked components to the \code{dashboard} object: A scatterplot, showing samples in along two-dimensional coordinates, and a violin plot, showing feature expression values by groups defined in \code{group_by}.
}
