% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/feature_grid_page.R
\name{add_feature_grid_page}
\alias{add_feature_grid_page}
\alias{add_feature_grid_page,i2dashboard,missing-method}
\alias{add_feature_grid_page,i2dashboard,SingleCellExperiment-method}
\alias{add_feature_grid_page,i2dashboard,Seurat-method}
\title{Create expression visualization for multiple selected features on a regular grid.}
\usage{
add_feature_grid_page(dashboard, object, ...)

\S4method{add_feature_grid_page}{i2dashboard,missing}(
  dashboard,
  use_dimred,
  exprs_values,
  title = "Feature grid",
  page = "feature_grid_page",
  menu = "Tools"
)

\S4method{add_feature_grid_page}{i2dashboard,SingleCellExperiment}(
  dashboard,
  object,
  use_dimred,
  exprs_values,
  subset_row = NULL,
  ...
)

\S4method{add_feature_grid_page}{i2dashboard,Seurat}(
  dashboard,
  object,
  use_dimred,
  assay,
  slot = "data",
  subset_row = NULL,
  ...
)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dashboard}.}

\item{object}{A \linkS4class{SingleCellExperiment} object or a \linkS4class{Seurat} object.}

\item{...}{Further parameters passed to the core function.}

\item{use_dimred}{Coordinates of the reduced dimensions, used for the scatterplot (see Details).}

\item{exprs_values}{Expression data of features of interest in rows and samples in columns (see Details).}

\item{title}{The title of the page.}

\item{page}{The name of the page to be added.}

\item{menu}{(Optional) The name of the menu, under which the page should appear.}

\item{subset_row}{A character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use. The default of NULL will use all features.}

\item{assay}{A character vector specifying which assay from \code{object@assays} to obtain expression values from.}

\item{slot}{A character vector specifying the name of the slot in the assay.}
}
\value{
An object of class \linkS4class{i2dashboard}.
}
\description{
Users can select features and a dimension reduction to plot feature expression values.
}
\details{
The parameters \code{use_dimred}, \code{exprs_values} (or \code{assay}) and \code{group_by} take different arguments depending on the class of \code{object}.
  In case no object is supplied (\emph{i2dashboard,missing}-method), the parameters are expected to be of class \code{data.frame} or \code{matrix}.
  In case a \linkS4class{SingleCellExperiment} or \linkS4class{Seurat} object is supplied, the parameters are expected to be of class \code{character}, containing
  \itemize{
    \item the name item(s) in \code{reducedDims(object)} or \code{object@reductions},
    \item a valid assay name from \code{assayNames(object)} or \code{names(object@assays)},
    \item column names of \code{colData(object)} or \code{object@meta.data}.
  }
}
