% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/heatmap.R
\name{heatmap}
\alias{heatmap}
\alias{heatmap,i2dashboard,missing-method}
\alias{heatmap,i2dashboard,SingleCellExperiment-method}
\alias{heatmap,i2dashboard,Seurat-method}
\title{Renders a component containing a \link[ComplexHeatmap]{Heatmap}.}
\usage{
heatmap(dashboard, object, ...)

\S4method{heatmap}{i2dashboard,missing}(
  dashboard,
  exprs_values,
  split_by = NULL,
  aggregate_by = NULL,
  title = NULL,
  legend = NULL,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  clustering_distance = c("euclidean", "maximum", "manhattan", "binary", "minkowski"),
  clustering_method = c("average", "ward.D", "ward.D2", "single", "complete",
    "mcquitty", "median", "centroid")
)

\S4method{heatmap}{i2dashboard,SingleCellExperiment}(
  dashboard,
  object,
  exprs_values = "counts",
  subset_row = NULL,
  split_by = NULL,
  aggregate_by = NULL,
  ...
)

\S4method{heatmap}{i2dashboard,Seurat}(
  dashboard,
  object,
  assay = "RNA",
  assay_slot = "data",
  subset_row = NULL,
  split_by = NULL,
  aggregate_by = NULL,
  ...
)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{A valid \linkS4class{SingleCellExperiment::SingleCellExperiment} object.}

\item{exprs_values}{A data.frame (matrix) containing expression data of features of interest in rows and samples in columns or a string representing the name of an \code{assay} of \code{object}.}

\item{split_by}{An optional data.frame (matrix) containing grouping factors for spliting columns of the heatmap. In case of \code{i2dashboard,SingleCellExperiment}, should be column names of \code{colData(object)}.}

\item{aggregate_by}{An optional data.frame (matrix) containing grouping factors for aggregating columns of the heatmap. In case of \code{i2dashboard,SingleCellExperiment}, should be column names of \code{colData(object)}.}

\item{title}{Title of the component.}

\item{legend}{Title of the heatmap legend.}

\item{cluster_rows}{A logical controls whether to make cluster on rows.}

\item{cluster_columns}{A logical controls whether to make cluster on columns.}

\item{clustering_distance}{The distance measure to use for hierarchical clustering.}

\item{clustering_method}{Method to perform hierarchical clustering, passed to \link[stats]{hclust}.}

\item{subset_row}{An optional character vector (of feature names), a logical vector or numeric vector (of indices) specifying the features to use. The default of NULL will use the first 100 features.}
}
\value{
A string containing markdown code for the rendered component

An object of class \linkS4class{i2dash::i2dashboard}.

An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
Renders a component containing a \link[ComplexHeatmap]{Heatmap}.
}
