% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plotly_barplot}
\alias{plotly_barplot}
\title{Render a barplot with plotly.}
\usage{
plotly_barplot(
  ...,
  showlegend = NULL,
  x_group_by_title = NULL,
  y_group_by_title = NULL
)
}
\arguments{
\item{...}{these arguments are of either the form \code{value} or \code{tag = value} and should be valid for the \code{plotly::plot_ly()} method.}

\item{showlegend}{(Optional) Boolean value that describes if the legend should be shown.}

\item{title_x}{(Optional) A title that describes the observations.}

\item{title_group_by}{(Optional) A title that describes the grouping factor.}
}
\value{
An object of class \code{plotly}.
}
\description{
Render a barplot with plotly.
}
