% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plotly_barplot}
\alias{plotly_barplot}
\title{Render a bar plot with plotly.}
\usage{
plotly_barplot(
  ...,
  showlegend = NULL,
  x_group_by_title = NULL,
  y_group_by_title = NULL
)
}
\arguments{
\item{...}{Additional arguments passed on to \code{plotly::plot_ly}.}

\item{showlegend}{(Optional) Boolean value that describes if the legend should be shown.}

\item{title_x}{(Optional) A title that describes the observations.}

\item{title_group_by}{(Optional) A title that describes the grouping factor.}
}
\value{
An object of class \code{plotly}.
}
\description{
Render a bar plot with plotly.
}
