% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plotly_barplot}
\alias{plotly_barplot}
\title{Render a bar plot with plotly.}
\usage{
plotly_barplot(
  ...,
  showlegend = NULL,
  x_group_by_title = NULL,
  y_group_by_title = NULL
)
}
\arguments{
\item{...}{Additional arguments passed on to \code{plotly::plot_ly}.}

\item{showlegend}{(Optional) Boolean value that describes if the legend should be shown.}

\item{x_group_by_title}{(Optional) A title that describes the grouping factor along the x-axis.}

\item{y_group_by_title}{(Optional) A title that describes the grouping factor along the y-axis.}
}
\value{
An object of class \code{plotly}.
}
\description{
Render a bar plot with plotly.
}
