% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plotly_boxplot}
\alias{plotly_boxplot}
\title{Render a boxplot with plotly.}
\usage{
plotly_boxplot(x, group_by = NULL, x_title = NULL,
  group_by_title = NULL, ...)
}
\arguments{
\item{x}{Numeric observations for the boxplot.}

\item{group_by}{A factor, by which observations can optionally be grouped.}

\item{x_title}{A title that describes the observations.}

\item{group_by_title}{A title that describes the grouping factor.}

\item{...}{these arguments are of the form tag = value and should be valid for the 'plotly::plot_ly()' method.}
}
\value{
An object of class \code{plotly}.
}
\description{
Render a boxplot with plotly.
}
