% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plotly_boxplot}
\alias{plotly_boxplot}
\title{Render a box plot with plotly.}
\usage{
plotly_boxplot(df, ..., title_x = NULL, title_group_by = NULL)
}
\arguments{
\item{df}{A dataframe containing the data fo the boxplot.}

\item{title_x}{A title that describes the observations.}

\item{title_group_by}{A title that describes the grouping factor.}

\item{x}{A named list with the observations for the boxplot.}

\item{group_by}{A list with factorial values, by which observations can optionally be grouped.}
}
\value{
An object of class \code{plotly}.
}
\description{
Render a box plot with plotly.
}
