% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plotly_scatterplot}
\alias{plotly_scatterplot}
\title{Creates a plotly scatterplot}
\usage{
plotly_scatterplot(..., y_title = NULL, x_title = NULL)
}
\arguments{
\item{df}{A dataframe containing the data for the boxplot}

\item{labels}{A list with sample names, which should be of the same length as x and y.}

\item{colour_by}{A list containing factorial values that will be used for colouring. In case of a named list, a dropdown menu will be provided in the interactive mode. Note: The length of all vectors in case of a named list should be of the same length as x and y.}

\item{checkbox}{A boolean value as indicator for colouring by labels.}

\item{selected_label}{The label (character) selected by the user.}
}
\value{
An object of class \code{plotly}.
}
\description{
Creates a plotly scatterplot
}
