% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plotly_scatterplot}
\alias{plotly_scatterplot}
\title{Render a barplot with plotly.}
\usage{
plotly_scatterplot(..., y_title = NULL, x_title = NULL)
}
\arguments{
\item{...}{Additional arguments passed on to \code{plotly::plot_ly}.}

\item{y_title}{The title of the x-axis.}

\item{x_title}{The title of the y-axis.}
}
\value{
An object of class \code{plotly}.
}
\description{
Render a barplot with plotly.
}
