% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plotly_scatterplot}
\alias{plotly_scatterplot}
\title{Renders a scatter plot}
\usage{
plotly_scatterplot(df, labels = NULL, colour_by = NULL,
  checkbox = FALSE, selected_label = NULL)
}
\arguments{
\item{labels}{A list with sample names, that should be of the same length as x and y.}

\item{colour_by}{A factor that will be mapped to colours. In case of a nested list, a dropdown menu will be provided in the interactive mode.}

\item{x}{Numeric values mapped to the x-axis. In case of a nested list, a dropdown menu will be provided in the interactive mode.}

\item{y}{Numeric values mapped to the y-axis. In case of a nested list, a dropdown menu will be provided in the interactive mode.}
}
\value{
A list with 1. the plotly object & 2. the data frame used in the plot
}
\description{
Renders a scatter plot
}
