% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plotly_violinplot}
\alias{plotly_violinplot}
\title{Render a vertical violin plot with plotly.}
\usage{
plotly_violinplot(df)
}
\arguments{
\item{df}{A data.frame containing in the first column numerical values for the y axis and optionally factor in the second column for grouping the observations along the x axis.}
}
\value{
An object of class \code{plotly}.
}
\description{
Render a vertical violin plot with plotly.
}
