% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/scatterplot.R
\docType{methods}
\name{scatterplot}
\alias{scatterplot}
\alias{scatterplot,i2dashboard,missing-method}
\alias{scatterplot,i2dashboard,SingleCellExperiment-method}
\alias{scatterplot,i2dashboard,Seurat-method}
\title{Renders a component containing a scatterplot with optional selection options}
\usage{
scatterplot(dashboard, object, ...)

\S4method{scatterplot}{i2dashboard,missing}(dashboard, x, y,
  colour_by = NULL, labels = NULL, exprs_values = NULL,
  title = NULL, x_title = NULL, y_title = NULL)

\S4method{scatterplot}{i2dashboard,SingleCellExperiment}(dashboard, object,
  use = c("colData", "rowData", "reducedDim"), x = NULL, y = NULL,
  colour_by = NULL, use_dimred = NULL, ...)

\S4method{scatterplot}{i2dashboard,Seurat}(dashboard, object,
  use = c("meta.data", "meta.features", "reduction"), x = NULL,
  y = NULL, colour_by = NULL, use_dimred = NULL, assay = "RNA",
  ...)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{An object of class \linkS4class{Seurat::Seurat} or \linkS4class{SingleCellExperiment::SingleCellExperiment}.}

\item{x}{A data.frame (matrix) containing columns with numeric values that will be mapped to the x-axis.}

\item{y}{A data.frame (matrix) containing columns with numeric values that will be mapped to the y-axis.}

\item{colour_by}{An optional data.frame (matrix) containing columns with numeric or factorial values that will be used for colouring.}

\item{labels}{An optional vector with sample names. A dropdown menu for colouring by label will be provided.}

\item{exprs_values}{An optional data.frame (matrix) containing expression data of features of interest in rows and samples in columns.}

\item{title}{The title of the components junk.}

\item{x_title}{An optional title of the x-axis. If not provided the column names from \code{x} are used instead.}

\item{y_title}{An optional title of the y-axis. If not provided the column names from \code{y}  are used instead.}

\item{use}{A character specifying where to obtain the data from. Valid input for SingleCellExperiment object: ("colData", "rowData", "reducedDim"). Valid input for Seurat object: ("meta.data" for sample metadata, "meta.feature" for feature metadata, "reduction" for a dimension reduction)}

\item{use_dimred}{A character vector indicating the reduced dimension to use from \code{"object"}.}

\item{assay}{Necessery, if \code{use} = "meta.feature". A character defining the assay to obtain the feature metadata from (default "RNA").}
}
\value{
A string containing markdown code for the rendered component

An object of class \linkS4class{i2dash::i2dashboard}.

An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
Renders a component containing a scatterplot with optional selection options
}
