% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Renders a Sequence saturation plot}
\usage{
scatterplot(object, x, y, colour_by = NULL, labels = NULL,
  title = NULL)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{x}{Numeric values mapped to the x-axis. In case of a nested list, a dropdown menu will be provided in the interactive mode.}

\item{y}{Numeric values mapped to the y-axis. In case of a nested list, a dropdown menu will be provided in the interactive mode.}

\item{colour_by}{A factor that will be mapped to colours. In case of a nested list, a dropdown menu will be provided in the interactive mode.}

\item{labels}{A list with sample names It should be of the same length as x and y. A dropdown menu for colouring by label will be provided.}

\item{title}{A title that will be displayed on top.}
}
\value{
A string containing markdown code for the rendered textbox
}
\description{
Renders a Sequence saturation plot
}
