% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/scatterplot.R
\docType{methods}
\name{scatterplot}
\alias{scatterplot}
\alias{scatterplot,i2dashboard,missing-method}
\alias{scatterplot,i2dashboard,SingleCellExperiment-method}
\title{Renders a component containing a scatterplot with optional selection options}
\usage{
scatterplot(dashboard, object, ...)

\S4method{scatterplot}{i2dashboard,missing}(dashboard, x, y,
  colour_by = NULL, labels = NULL, exprs_values = NULL,
  title = NULL, x_title = NULL, y_title = NULL)

\S4method{scatterplot}{i2dashboard,SingleCellExperiment}(dashboard, object,
  use = "colData", x = NULL, y = NULL, colour_by = NULL,
  reduced_dim = NULL, ...)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{A valid \linkS4class{SingleCellExperiment::SingleCellExperiment} object.}

\item{x}{A data.frame (matrix) containing columns with numeric values that will be mapped to the x-axis.}

\item{y}{A data.frame (matrix) containing columns with numeric values that will be mapped to the y-axis.}

\item{colour_by}{An optional data.frame (matrix) containing columns with numeric or factorial values that will be used for colouring.}

\item{labels}{An optional vector with sample names. A dropdown menu for colouring by label will be provided.}

\item{exprs_values}{An optional data.frame (matrix) containing expression data of features of interest in rows and samples in columns.}

\item{title}{The title of the components junk.}

\item{x_title}{An optional title of the x-axis. If not provided the column names from \code{x} are used instead.}

\item{y_title}{An optional title of the y-axis. If not provided the column names from \code{y}  are used instead.}

\item{use}{A character specifying where to obtain the data from. One of \code{"colData"}, \code{"rowData"}, \code{"reducedDim"}.}

\item{reducedDim}{A character vector indicating the reduced dimension to use from \code{"reducedDim"}}
}
\value{
A string containing markdown code for the rendered component

An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
Renders a component containing a scatterplot with optional selection options
}
