% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_saturation.R
\name{sequence_saturation}
\alias{sequence_saturation}
\title{Renders a Sequence saturation plot}
\usage{
sequence_saturation(object, x, y, colour_by = NULL,
  title = "Sequencing saturation")
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{x}{Numeric values mapped to the x-axis. In case of a nested list, a dropdown menu will be provided in the interactive mode.}

\item{y}{Numeric values mapped to the y-axis. In case of a nested list, a dropdown menu will be provided in the interactive mode.}

\item{title}{A title that will be displayed on top.}

\item{coulor_by}{A factor that will be mapped to colours.In case of a nested list, a dropdown menu will be provided in the interactive mode.}
}
\value{
A string containing markdown code for the rendered textbox
}
\description{
Renders a Sequence saturation plot
}
