% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\alias{summarize_samples}
\alias{summarize_features}
\title{Summarize samples or features of objects containing data from single-cell experiments.}
\usage{
summarize_samples(object, ...)

summarize_features(object, ...)
}
\arguments{
\item{object}{A SingleCellExperiment or Seurat object.}

\item{columns}{The metadata columns to summarize.}

\item{FUNS}{A named vector indicating summary functions.}
}
\value{
A \code{data.frame} with one row for each \code{columns} and one column for each \code{FUNS}, containing the desired summary. An additional column \code{group} is added which indicates whether values from \emph{Samples} or \emph{Features} were summarized.
}
\description{
Summarize samples or features of objects containing data from single-cell experiments.
}
