% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_data_}
\alias{summarize_data_}
\title{Summarizes columns of \code{data} using \code{FUNS} and can be grouped by \code{group_by}.}
\usage{
summarize_data_(data, columns, FUNS, group_by = NULL)
}
\arguments{
\item{data}{A data.frame.}

\item{columns}{The columns of \code{data} that are summarized.}

\item{FUNS}{A named vector of summary functions}

\item{group_by}{A vector of the same length as the number of rows in data, which is used for grouping.}
}
\value{
A \code{data.frame} that contains summary statistics.
}
\description{
Summarizes columns of \code{data} using \code{FUNS} and can be grouped by \code{group_by}.
}
