% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/summarize.R
\name{summarize_metadata}
\alias{summarize_metadata}
\alias{summarize_metadata,i2dashboard,missing-method}
\alias{summarize_metadata,i2dashboard,SingleCellExperiment-method}
\alias{summarize_metadata,i2dashboard,Seurat-method}
\title{Renders a component containing an HTML table of summarized sample or feature metadata of single-cell experiment objects or data.frames.}
\usage{
summarize_metadata(dashboard, object, ...)

\S4method{summarize_metadata}{i2dashboard,missing}(
  dashboard,
  data,
  columns = colnames(data),
  FUNS = c("min", "max", "mean", "median"),
  group_by = NULL,
  caption = NULL,
  description = NULL,
  title = NULL
)

\S4method{summarize_metadata}{i2dashboard,SingleCellExperiment}(
  dashboard,
  object,
  from = c("colData", "rowData"),
  columns = NULL,
  group_by = NULL,
  caption = "Table: ",
  description = NULL,
  ...
)

\S4method{summarize_metadata}{i2dashboard,Seurat}(
  dashboard,
  object,
  from = c("meta.data", "meta.features"),
  assay = "RNA",
  columns = NULL,
  group_by = NULL,
  caption = "Table",
  description = NULL,
  ...
)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dashboard}.}

\item{object}{An object of class \linkS4class{Seurat} or \linkS4class{SingleCellExperiment}.}

\item{...}{Further parameters passed to the core function.}

\item{data}{Data.frame containing numeric observations to summarize.}

\item{columns}{The metadata columns to summarize.}

\item{FUNS}{A named vector indicating summary functions.}

\item{group_by}{A column name of a column containing grouping factors for grouping.}

\item{caption}{The tables caption.}

\item{description}{The tables description.}

\item{title}{The title of the component.}

\item{from}{A character specifying where to obtain the data from \code{object}, e.g. \code{colData}/\code{rowData}.}

\item{assay}{A character defining the assay of \code{object}.}
}
\value{
A string containing markdown code for the rendered component
}
\description{
Renders a component containing an HTML table of summarized sample or feature metadata of single-cell experiment objects or data.frames.
}
