% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinplot.R
\docType{methods}
\name{violinplot}
\alias{violinplot}
\alias{violinplot,i2dashboard,missing-method}
\alias{violinplot,i2dashboard,SingleCellExperiment-method}
\title{Renders a component containing a vertical violin plot}
\usage{
violinplot(object, sc_object, ...)

\S4method{violinplot}{i2dashboard,missing}(object, y, group_by = NULL,
  title = NULL)

\S4method{violinplot}{i2dashboard,SingleCellExperiment}(object, sc_object,
  use = "colData", y = NULL, group_by = NULL, title = NULL)
}
\arguments{
\item{object}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{sc_object}{A valid \linkS4class{SingleCellExperiment::SingleCellExperiment} object.}

\item{y}{A data.frame or matrix with columns containing numeric observations for the vertical violinplot (y-axis).}

\item{group_by}{(Optional) A data.frame or matrix with columns conatining factors for grouping the observations.}

\item{title}{(Optional) The title of the components junk.}

\item{use}{A character specifying where to obtain the data from. Valid input: "colData" or "rowData".}

\item{y}{A character vector with column names to use for observations.}

\item{group_by}{A character vector with column names to use for grouping the observations.}
}
\value{
A string containing markdown code for the rendered component

An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
Renders a component containing a vertical violin plot
}
