% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinplot.R
\name{violinplot}
\alias{violinplot}
\title{Renders a vertical violin plot}
\usage{
violinplot(object, y, group_by = NULL, title = NULL, title_y = NULL,
  title_group_by = NULL)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{y}{A vector with numerical values or a named list, which represents the observations for the vertical violinplot (y-axis). In case of a named list, a dropdown menu will be provided in the interactive mode.}

\item{group_by}{(Optional) A vector with factorial values or characters or a named list, which will be used for grouping the observations. In case of a named list, a dropdown menu will be provided in the interactive mode.}

\item{title}{(Optional) The title of the components junk.}

\item{title_y}{(Optional) The title of the y-axis that describes the observations. In case of a named list this parameter is not needed because the names of the list will be used as title of the y axis.}

\item{title_group_by}{(Optional) The title of the x-axis that describes the grouping factor. In case of a named list this parameter is not needed because the names of the list will be used as title of the x axis.}
}
\value{
A string containing markdown code for the rendered component
}
\description{
Renders a vertical violin plot
}
