% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/violinplot.R
\docType{methods}
\name{violinplot}
\alias{violinplot}
\alias{violinplot,i2dashboard,missing-method}
\alias{violinplot,i2dashboard,SingleCellExperiment-method}
\title{Renders a component containing a vertical violin plot}
\usage{
violinplot(dashboard, object, ...)

\S4method{violinplot}{i2dashboard,missing}(dashboard, y, group_by = NULL,
  title = NULL, y_title = NULL, group_by_title = NULL)

\S4method{violinplot}{i2dashboard,SingleCellExperiment}(dashboard, object,
  use = "colData", y = NULL, group_by = NULL, title = NULL,
  y_title = NULL, group_by_title = NULL)
}
\arguments{
\item{dashboard}{An object of class \linkS4class{i2dash::i2dashboard}.}

\item{object}{A valid \linkS4class{SingleCellExperiment::SingleCellExperiment} object.}

\item{y}{A data.frame (matrix) containing numeric observations for the vertical axis, or a character vector indicating column names of \code{colData(object)}, \code{rowData(object)}.}

\item{group_by}{An optional data.frame (matrix) with columns containing grouping factors for the horizontal axis.}

\item{title}{The title of the component.}

\item{y_title}{The title of the y-axis.}

\item{group_by_title}{The title of the x-axis.}

\item{use}{A character specifying where to obtain the data from. One of \code{"colData"} or \code{"rowData"}.}
}
\value{
A string containing markdown code for the rendered component

An object of class \linkS4class{i2dash::i2dashboard}.
}
\description{
Renders a component containing a vertical violin plot
}
